### 2023-03-16 #5264 @danielskatz
👋 @paulbrodersen - thanks for your submission.  There are a couple of issues:

1. Note that your paper does not compile. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) and note that you can click on the error above to find out more about it. In addition, you could work on the possibly missing DOIs that editorialbot suggests, but note that some may be incorrect. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` after making changes to the .md file or when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

2. I'm going to start a scope review amongst the JOSS editors to decide if this submission is [research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software) as defined by JOSS. This does not mean that it is not software that is useful in research, but just that JOSS might not consider it in scope for review as research software. In particular, visualization software that does not also do some analysis is often considered out of scope.
