+++ publication/paper.bib
@@ -33,7 +33,8 @@
   publisher ="Springer Berlin Heidelberg",
   address   ="Berlin, Heidelberg",
   pages     ="483--484",
-  isbn      ="978-3-540-45848-7"
+  isbn      ="978-3-540-45848-7",
+  doi       ="10.1007/3-540-45848-4_57"
   }
 @article{Shannon:2003,
   author   = {Shannon, Paul and Markiel, Andrew and Ozier, Owen and Baliga, Nitin S. and Wang, Jonathan T. and Ramage, Daniel and Amin, Nada and Schwikowski, Benno and Ideker, Trey},
@@ -56,7 +57,8 @@ author     = {Mathieu Bastian and Sebastien Heymann and Mathieu Jacomy},
 title      = {Gephi: An Open Source Software for Exploring and Manipulating Networks},
 conference = {International AAAI Conference on Weblogs and Social Media},
 year       = {2009},
-url        = {http://www.aaai.org/ocs/index.php/ICWSM/09/paper/view/154}
+url        = {http://www.aaai.org/ocs/index.php/ICWSM/09/paper/view/154},
+doi        = {10.1609/icwsm.v3i1.13937}
 }
 @article{Hunter:2007,
   Author    = {Hunter, J. D.},
