### 2023-03-16 #5264 @paulbrodersen
Regarding potential reviewers, my preferred choices would be **tacaswell** or **briatte**. Netgraph builds on matplotlib, and tacaswell (matplotlib's lead maintainer) is aware of my work as he has helped me with particularly gnarly issues in the past. briatte has authored a very similar library (ggnetwork) for the R programming language, and would hence be in a unique position to judge the merits this work.

Based on the linked table, the following people also seem highly suitable given that their preferred programming language includes python, and their domains of expertise include network analysis or network science and visualisation. However, I am unfamiliar with their work (or at least their github handle).

- jankatins
- alexanderfurnas
- adavidzh
- camillescott
- ortega2247
- tomalrussell
- jkbren
