### 2023-03-20 184209b
Add support for empty graphs in `EditableGraph` (issue #62).
