### 2023-03-20 2d35c39
Add support for graphs with nodes but no edges to `EditableGraph` (issue #62)
