+++ netgraph/_parser.py
@@ -44,8 +44,7 @@ def _is_listlike(graph):
 @_handle_multigraphs
 def _parse_sparse_matrix_format(adjacency):
     """Parse graphs given in a sparse format, i.e. edge lists or sparse matrix representations."""
-    adjacency = np.array(adjacency)
-    rows, columns = adjacency.shape
+    rows, columns = np.array(adjacency).shape
 
     if columns == 2:
         edges = _parse_edge_list(adjacency)
@@ -53,9 +52,9 @@ def _parse_sparse_matrix_format(adjacency):
         return nodes, edges, None
 
     elif columns == 3:
-        edges = _parse_edge_list(adjacency[:, :2])
-        nodes = _get_unique_nodes(edges)
         edge_weight = {(source, target) : weight for (source, target, weight) in adjacency}
+        edges = list(edge_weight.keys())
+        nodes = _get_unique_nodes(edges)
 
         # In a sparse adjacency format with integer nodes and float weights,
         # the type of nodes is promoted to the same type as weights.
@@ -92,17 +91,17 @@ def _is_nparray(graph):
 
 
 def _parse_nparray(graph):
-        rows, columns = graph.shape
-        if columns in (2, 3):
-            return _parse_sparse_matrix_format(graph)
-        elif rows == columns:
-            return _parse_adjacency_matrix(graph)
-        else:
-            msg = "Could not interpret input graph."
-            msg += "\nIf a graph is specified as a numpy array, it has to have one of the following shapes:"
-            msg += "\n\t-(E, 2) or (E, 3), where E is the number of edges"
-            msg += "\n\t-(V, V), where V is the number of nodes (i.e. full rank)"
-            msg += f"However, the given graph had shape {graph.shape}."
+    rows, columns = graph.shape
+    if columns in (2, 3):
+        return _parse_sparse_matrix_format(graph)
+    elif rows == columns:
+        return _parse_adjacency_matrix(graph)
+    else:
+        msg = "Could not interpret input graph."
+        msg += "\nIf a graph is specified as a numpy array, it has to have one of the following shapes:"
+        msg += "\n\t-(E, 2) or (E, 3), where E is the number of edges"
+        msg += "\n\t-(V, V), where V is the number of nodes (i.e. full rank)"
+        msg += f"However, the given graph had shape {graph.shape}."
 
 
 def _parse_adjacency_matrix(adjacency):
@@ -222,26 +221,39 @@ def parse_graph(graph):
         except ModuleNotFoundError:
             pass
     else:
-        allowed = ['list', 'tuple', 'set', 'networkx.Graph', 'igraph.Graph', 'graphtool.Graph']
+        allowed = ['list', 'tuple', 'set', 'networkx.Graph', 'igraph.Graph', 'graph_tool.Graph']
         raise NotImplementedError("Input graph must be one of: {}\nCurrently, type(graph) = {}".format("\n\n\t" + "\n\t".join(allowed), type(graph)))
 
 
-_check_to_is_empty = {
-    _is_listlike   : lambda x : not(x),
-    _is_nparray    : lambda x : False if x.any() else True, # not quite correct, as np.array([0, 0]) and np.array([0, 0, 0]) would indicate self-loops around node 0
-    _is_networkx   : lambda x : not(x),
-    _is_igraph     : lambda x : not(x),
-    _is_graph_tool : lambda x : not(x.get_vertices().tolist()),
-}
-
-
-def is_empty(graph):
-    for check, _is_empty in _check_to_is_empty.items():
+def is_order_zero(graph):
+    """Determine if a graph is an order zero graph, i.e. a graph with no nodes (and no edges)."""
+    for check, parser in _check_to_parser.items():
         try:
             if check(graph):
-                return _is_empty(graph)
[…]
