### 2023-03-21 7d27672
Bugfix: rectangle selector now works with recent matplotlib versions
