### 2023-03-22 6b3c877
Bugfix: Convert ragged array `node_spacings` in `get_multipartite_layout` to list (issue #65)
Ragged arrays are disallowed in recent numpy versions.
