+++ setup.py
@@ -6,7 +6,7 @@ def read_file(filename):
     with open(os.path.join(os.path.dirname(__file__), filename)) as file:
         return file.read()
 
-version = '4.12.7'
+version = '4.12.8'
 
 setup(
     name='netgraph',
