### 2023-03-29 c883513
Bugfix: handle distances below or equal to zero in `_shorten_line` correctly
