### 2023-03-29 e54e314
Bugfix: disable automatic node label font size rescaling if the fontsize is specified using the `fontsize` keyword
Previously, only the `size` keyword resulted in the correct
behaviour.
