### 2023-03-29 e7f37c6
Support arbitrary node shapes specified by `matplotlib.path.Path` instances
