+++ netgraph/_artists.py
@@ -52,79 +52,120 @@ class PathPatchDataUnits(PathPatch):
 
 
 class NodeArtist(PathPatchDataUnits):
-    """Implements the node artists class.
+    """NodeArtist base class.
 
     Parameters
     ----------
-    shape : str
-        The shape of the node. Specification is as for matplotlib.scatter marker, i.e. one of 'so^>v<dph8'.
+    shape : str or matplotlib.Path instance
+        The shape of the node.
+        If a string, the specification is as for matplotlib.scatter marker, i.e. one of 'so^>v<dph8'.
+        If a path, the path has to be closed and the vertices have to be centered on (0,0).
     xy : tuple
         The (float x, float y) coordinates of the centroid.
-    radius : float
+    size : float
         The distance from the center to each of the vertices.
     **kwargs
         `Patch` properties:
         %(Patch_kwdoc)s
 
+    See also
+    --------
+    CircularNodeArtist, RegularPolygonArtist
+
     """
 
-    def __init__(self, shape, xy, radius, **kwargs):
-        self.shape = shape
+    def __init__(self, path, xy, size, orientation=0, linewidth_correction=2, **kwargs):
+        self._path = path
         self.xy = xy
-        self.radius = radius
-
-        if shape == 'o': # circle
-            self.numVertices = None
-            self.orientation = 0
-        elif shape == '^': # triangle up
-            self.numVertices = 3
-            self.orientation = 0
-        elif shape == '<': # triangle left
-            self.numVertices = 3
-            self.orientation = np.pi*0.5
-        elif shape == 'v': # triangle down
-            self.numVertices = 3
-            self.orientation = np.pi
-        elif shape == '>': # triangle right
-            self.numVertices = 3
-            self.orientation = np.pi*1.5
-        elif shape == 's': # square
-            self.numVertices = 4
-            self.orientation = np.pi*0.25
-        elif shape == 'd': # diamond
-            self.numVertices = 4
-            self.orientation = np.pi*0.5
-        elif shape == 'p': # pentagon
-            self.numVertices = 5
-            self.orientation = 0
-        elif shape == 'h': # hexagon
-            self.numVertices = 6
-            self.orientation = 0
-        elif shape == '8': # octagon
-            self.numVertices = 8
-            self.orientation = 0
-        else:
-            raise ValueError("Node shape should be one of: 'so^>v<dph8'. Current shape:{}".format(shape))
-
-        if self.shape == 'o': # circle
-            self.linewidth_correction = 2
-            self._path = Path.circle()
-        else: # regular polygon
-            self.linewidth_correction = 2 * np.sin(np.pi/self.numVertices) # derives from the ratio between a side and the radius in a regular polygon.
-            self._path = Path.unit_regular_polygon(self.numVertices)
-
+        self.size = size
+        self.orientation = orientation
+        self.linewidth_correction = linewidth_correction
         self._patch_transform = transforms.Affine2D()
         super().__init__(path=self._path, **kwargs)
 
-    def get_path(self):
-        return self._path
-
     def get_patch_transform(self):
-        # The factor 2 * sin(pi/n) d
         return self._patch_transform.clear() \
-            .scale(self.radius-self._lw_data/self.linewidth_correction) \
+            .scale(self.size-self._lw_data/self.linewidth_correction) \
             .rotate(self.orientation) \
             .translate(*self.xy)
+
+    def get_head_offset(self, edge_path):
+        # Determine edge path points that are within node shape path.
+        is_overlapping = self._path.contains_points(edge_path, transform=self.get_patch_transform())
+
[…]
