+++ tests/toy_graphs.py
@@ -21,7 +21,9 @@ cycle = [
     (4, 5),
     (5, 6),
     (6, 7),
-    (7, 0)
+    (7, 8),
+    (8, 9),
+    (9, 0)
 ]
 
 triangle = [(0, 1), (1, 2), (2, 0)]
