### 2023-03-31 0f07234
Support node label rescaling for arbitrary node shapes
