### 2023-04-07 #5264 @rkurchin
@adavidzh, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

(@danielskatz, sorry for delay here – my wedding was this past weekend, so there were a few other things on my mind 🤪 )
