### 2023-04-08 #5264 @adavidzh
> @adavidzh, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

I am not able to commit the time to take this through right now. I did take a cursory look at the repository and I could not identify community guidelines.

> (@danielskatz, sorry for delay here – my wedding was this past weekend, so there were a few other things on my mind 🤪 )

Congratulations!
