### 2023-04-10 #5264 @jonjoncardoso
Hi,

I am still happy to do the review. However, I underestimated the time I’d
have before the Easter Break. I’m travelling right now and I’ll only be
able to do so in early may. If that’s not a problem, or if you can’t find
alternative reviewers, I can still review by then.

On Sun, 9 Apr 2023 at 16:23, Rachel Kurchin ***@***.***>
wrote:

> 👋 @ortega2247 <https://github.com/ortega2247> and/or @tomalrussell
> <https://github.com/tomalrussell>, would you be willing to review this
> submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
