### 2023-04-11 #5264 @paulbrodersen
@adavidzh  Hi, thanks for checking out my project. You are correct that I don't have a section that is titled "Community Guidelines". I do, however, have the following paragraphs that serve the same purpose under the heading "Contributing & Support" on ReadTheDocs and under "Help, I don't know how to do ..." on the Github README:

---

*If you get stuck, please raise an issue on [Github](https://github.com/paulbrodersen/netgraph/issues). Include any relevant code and data in a [minimal, reproducible example](https://stackoverflow.com/help/minimal-reproducible-example). If applicable, make a sketch of the desired result with pen and paper, take a picture, and append it to the issue.
Bug reports are, of course, always welcome. Please make sure to include the full error trace.
If you submit a pull request that fixes a bug or implements a cool feature, I will probably worship the ground you walk on for the rest of the week. Probably.*

---

If there is anything missing to qualify as "Community Guidelines" or if you have any other suggestions for improvements, then I would love to hear them.
