### 2023-04-12 #5264 @paulbrodersen
@rkurchin: Before @tomalrussell puts any work into this, I just wanted to double-check what JOSS' conflicts of interest policy says about the reviewer and reviewee being from the same university? @tomalrussell and I have never met (as far as I know) and we work in different departments but I noticed on his github profile that we are both at Oxford.
