### 2023-04-12 #5264 @tomalrussell
> double-check what JOSS' conflicts of interest policy says about the reviewer and reviewee being from the same university

Ah, good catch @paulbrodersen - https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy does flag this case. Hopefully this can be noted and waived as we are "both employed by the same very large organization but in different units without any knowledge of each other." (I'm over in Geography, and I don't believe we've met - yet!)
