### 2023-04-12 28e7351
Bugfix: handle self-loops correctly when computing edge head and tail offsets.
