### 2023-04-12 d8acaea
Bugfix: set axis aspect and update view before computing label fontsizes
