### 2023-04-12 4b684b8
Remove superfluous _update_view method
