### 2023-04-13 18f6694
Move normalisation & rescaling functions to _utils.py
