+++ netgraph/_main.py
@@ -22,6 +22,11 @@ from ._utils import (
     _rank,
     _get_n_points_on_a_circle,
     _edge_list_to_adjacency_list,
+    _normalize_numeric_argument,
+    _normalize_color_argument,
+    _normalize_shape_argument,
+    _rescale_dict_values,
+    _check_completeness,
 )
 from ._node_layout import (
     get_fruchterman_reingold_layout,
@@ -291,22 +296,22 @@ class BaseGraph(object):
         self.nodes = self._initialize_nodes(nodes)
 
         # Convert all node and edge parameters to dictionaries.
-        node_shape      = self._normalize_shape_argument(node_shape, self.nodes, 'node_shape')
-        node_size       = self._normalize_numeric_argument(node_size, self.nodes, 'node_size')
-        node_edge_width = self._normalize_numeric_argument(node_edge_width, self.nodes, 'node_edge_width')
-        node_color      = self._normalize_color_argument(node_color, self.nodes, 'node_color')
-        node_edge_color = self._normalize_color_argument(node_edge_color, self.nodes, 'node_edge_color')
-        node_alpha      = self._normalize_numeric_argument(node_alpha, self.nodes, 'node_alpha')
-        node_zorder     = self._normalize_numeric_argument(node_zorder, self.nodes, 'node_zorder')
-        edge_width      = self._normalize_numeric_argument(edge_width, self.edges, 'edge_width')
-        edge_color      = self._normalize_color_argument(edge_color, self.edges, 'edge_color')
-        edge_alpha      = self._normalize_numeric_argument(edge_alpha, self.edges, 'edge_alpha')
-        edge_zorder     = self._normalize_numeric_argument(edge_zorder, self.edges, 'edge_zorder')
+        node_shape      = _normalize_shape_argument(node_shape, self.nodes, 'node_shape')
+        node_size       = _normalize_numeric_argument(node_size, self.nodes, 'node_size')
+        node_edge_width = _normalize_numeric_argument(node_edge_width, self.nodes, 'node_edge_width')
+        node_color      = _normalize_color_argument(node_color, self.nodes, 'node_color')
+        node_edge_color = _normalize_color_argument(node_edge_color, self.nodes, 'node_edge_color')
+        node_alpha      = _normalize_numeric_argument(node_alpha, self.nodes, 'node_alpha')
+        node_zorder     = _normalize_numeric_argument(node_zorder, self.nodes, 'node_zorder')
+        edge_width      = _normalize_numeric_argument(edge_width, self.edges, 'edge_width')
+        edge_color      = _normalize_color_argument(edge_color, self.edges, 'edge_color')
+        edge_alpha      = _normalize_numeric_argument(edge_alpha, self.edges, 'edge_alpha')
+        edge_zorder     = _normalize_numeric_argument(edge_zorder, self.edges, 'edge_zorder')
 
         # Rescale.
-        node_size = self._rescale(node_size, BASE_SCALE)
-        node_edge_width = self._rescale(node_edge_width, BASE_SCALE)
-        edge_width = self._rescale(edge_width, BASE_SCALE)
+        node_size       = _rescale_dict_values(node_size, BASE_SCALE)
+        node_edge_width = _rescale_dict_values(node_edge_width, BASE_SCALE)
+        edge_width      = _rescale_dict_values(edge_width, BASE_SCALE)
 
         self.node_size = node_size
 
@@ -377,92 +382,6 @@ class BaseGraph(object):
                 raise ValueError(msg)
 
 
-    def _normalize_numeric_argument(self, numeric_or_dict, dict_keys, variable_name):
-        if isinstance(numeric_or_dict, (int, float)):
-            return {key : numeric_or_dict for key in dict_keys}
[…]
