### 2023-04-15 #5372 @editorialbot
[REVIEW]: Netgraph: Publication-quality Network Visualisations in Python
**Submitting author:** @paulbrodersen (<a href="http://orcid.org/0000-0001-5216-7863">Paul Brodersen</a>)
**Repository:** https://github.com/paulbrodersen/netgraph/
**Branch with paper.md** (empty if default branch): joss
**Version:** 4.12.4
