### 2023-04-20 221f783
Support variable self-loop radii
