### 2023-04-20 925fc0d
Implement automatic scaling of unspecified self-loop radii
