### 2023-04-21 2f76f2a
Improve optimisation of node label rescaling
