### 2023-04-24 090099c
Remove _get_font_size method
