### 2023-04-24 7445638
Reduce node label re-rendering when auto-scaling the font size
