### 2023-04-24 0c70038
Remove superfluous _get_text_object_bbox and _get_text_object_dimensions
