### 2023-04-24 b121b2a
Make `rescale_node_labels` and `maximize_node_labels` private methdods
