### 2023-04-24 c2912e2
Re-scale node labels on resize if `autoscale_node_labels` is True
