### 2023-04-25 ae17c66
Move `_get_optimal_offsets` into _utils.py
