### 2023-04-25 0f666c9
Support variable self-loop angles
