+++ netgraph/_edge_layout.py
@@ -125,7 +125,7 @@ def _shift_edge(x1, y1, x2, y2, delta):
     return x1+dx, y1+dy, x2+dx, y2+dy
 
 
-def get_selfloop_paths(edges, node_positions, selfloop_radius, origin, scale, angle=None):
+def get_selfloop_paths(edges, node_positions, selfloop_radius, selfloop_angle, origin, scale):
     """Edge routing for self-loops.
 
     Parameters
@@ -136,12 +136,12 @@ def get_selfloop_paths(edges, node_positions, selfloop_radius, origin, scale, an
         Dictionary mapping each node ID to (float x, float y) tuple, the node position.
     selfloop_radius : dict or float, default 0.1
         Dictionary mapping each self-loop edge to a radius. If float, all self-loops have the same radius.
+    selfloop_angle : dict or float
+        The starting angle of the self-loop in radians.
     origin : numpy.array
         A (float x, float y) tuple corresponding to the lower left hand corner of the bounding box specifying the extent of the canvas.
     scale : numpy.array
         A (float x, float y) tuple representing the width and height of the bounding box specifying the extent of the canvas.
-    angle : float or None, default None
-        The starting angle of the self-loop in radians. If None, the self-loop is drawn opposite of the centroid of the graph.
 
     Returns
     -------
@@ -151,16 +151,17 @@ def get_selfloop_paths(edges, node_positions, selfloop_radius, origin, scale, an
     """
     selfloops = [(source, target) for (source, target) in edges if source == target]
     selfloop_radius = _normalize_numeric_argument(selfloop_radius, selfloops, 'selfloop_radius')
+    selfloop_angle = _normalize_numeric_argument(selfloop_angle, selfloops, 'angle', allow_none=True)
 
     edge_paths = dict()
     for source, target in selfloops:
         edge_paths[(source, target)] = _get_selfloop_path(
-            source, node_positions, selfloop_radius[(source, target)], origin, scale, angle)
+            source, node_positions, selfloop_radius[(source, target)], selfloop_angle[(source, target)], origin, scale)
 
     return edge_paths
 
 
-def _get_selfloop_path(source, node_positions, selfloop_radius, origin, scale, angle=None):
+def _get_selfloop_path(source, node_positions, radius, angle, origin, scale):
     """Compute the edge path for a single self-loop."""
 
     x, y = node_positions[source]
@@ -178,21 +179,22 @@ def _get_selfloop_path(source, node_positions, selfloop_radius, origin, scale, a
         else: # single node in graph; self-loop points upwards
             unit_vector = np.array([0, 1])
 
-    selfloop_center = node_positions[source] + selfloop_radius * unit_vector
+    center = node_positions[source] + radius * unit_vector
 
-    selfloop_path = _get_n_points_on_a_circle(
-        selfloop_center, selfloop_radius, 100+1,
+    path = _get_n_points_on_a_circle(
+        center, radius, 100+1,
         _get_angle(*unit_vector) + np.pi,
     )[1:]
 
     # # ensure that the loop stays within the bounding box
-    # selfloop_path = _clip_to_frame(selfloop_path, origin, scale)
+    # path = _clip_to_frame(path, origin, scale)
 
-    return selfloop_path
+    return path
 
 
 def get_curved_edge_paths(edges, node_positions,
                           selfloop_radius               = 0.1,
+                          selfloop_angle                = None,
                           origin                        = np.array([0, 0]),
                           scale                         = np.array([1, 1]),
                           k                             = None,
@@ -215,6 +217,8 @@ def get_curved_edge_paths(edges, node_positions,
         Dictionary mapping each node ID to (float x, float y) tuple, the node position.
[…]
