### 2023-04-26 da2241e
Remove unused `edge_width` parameter in `get_straight_edge_paths`
