### 2023-04-26 4f02143
Remove unused `origin` and `scale` parameters when computing self-loop paths
