### 2023-04-26 9c84b59
fixup! Remove unused `edge_width` parameter in `get_straight_edge_paths`
