### 2023-04-26 04cdef8
fixup! Remove unused `origin` and `scale` parameters when computing self-loop paths
