### 2023-04-26 2343e80
fixup! Remove unused `edge_width` parameter in `get_straight_edge_paths`
