### 2023-04-26 356548c
fixup! Remove unused `origin` and `scale` parameters when computing self-loop paths
