### 2023-04-26 4ec5c8d
Remove unused `origin` and `scale` parameters from `get_arced_edge_paths`
