### 2023-04-27 95290e2
Handle self-loops inside `get_straight_edge_paths` and `get_arced_edge_paths`
Previously, self-loop paths were computed independently.
