### 2023-05-02 dab8710
Resample edge paths more efficiently
