### 2023-05-02 684af3d
Improve readability of `_get_optimal_offsets`
