### 2023-05-02 b4d6ef6
Optimise self-loop angles in `get_straight_edge_paths`
