### 2023-05-03 816727b
Remove `get_selfloop_paths`
