### 2023-05-03 043279e
Simplify `get_straight_edge_paths`
