### 2023-05-03 876f1e7
Move `_shift_edge` to end of file; tag for removal
