### 2023-05-03 fa5e46f
Simplify code for initialising control points in `get_curved_edge_paths`
