### 2023-05-05 a9edeb4
Remove superfluous `k` parameter in `_initialize_control_points`
