### 2023-05-05 0e7b49e
Consistent variable names for edge end points
