### 2023-05-05 9cf87dc
Move `_expand_edges` definition
