### 2023-05-05 4f77ae5
Merge `_initialize_nonloops` and `_initialize_selfloops`
