### 2023-05-05 ad514e6
Optimise self-loop angles in `get_curved_edge_paths`
