### 2023-05-05 3a527be
Remove redundant setting of default values
