### 2023-05-05 0bb5235
Change defaults in `get_bundled_edges`
