+++ README.md
@@ -171,6 +171,7 @@ The following changes are worth noting:
 - The `selfloop_radius` and `selfloop_angle` arguments can now also be dictionaries that maps self-loop edges to radii/angles. Previously, only a single float argument was accepted that was applied to all self-loops.
 - `get_straight_edge_paths` and `get_arced_edge_paths` now also compute the paths for self-loops, and hence accept additional arguments. This change facilitated improvements to the placements of self-loops, reducing collisions with other nodes and edges. Previously, self-loops were computed independently.
 - Removed `get_selfloop_paths`, which is now superfluous.
+- In `get_bundled_edge_paths` the default values for `k` and `total_cycles` were changed to 500 (from 1000) and 6 (from 5), respectively, as these values typically seem to give better results.
 
 
 ## Help, I don't know how to do ...!
