### 2023-05-05 ff2f242
Set `k` parameter in `get_curved_edges` explicitly
