### 2023-05-05 3336140
fixup! Remove redundant setting of default values
