### 2023-05-11 #5372 @jonjoncardoso
Hi Rachel! Apologies for the long delay. I can still do it. Hopefully I can
send my notes by this weekend

On Thu, 11 May 2023 at 18:15, Rachel Kurchin ***@***.***>
wrote:

> @ortega2247 <https://github.com/ortega2247>, please let me know if you'll
> be able to finish review. @jonjoncardoso
> <https://github.com/jonjoncardoso>, if you can no longer review, please
> also let me know and I'll find an alternative reviewer.
>
