### 2023-05-14 paulbrodersen/netgraph#68 @ortega2247
Installation not working with python 3.11
I tried installing netgraph using conda and python 3.11 in a windows machine and got the following error:

`ImportError: DLL load failed while importing _imaging: The specified module could not be found.
`
