### 2023-05-14 #5372 @ortega2247
Minor issue: https://github.com/paulbrodersen/netgraph/issues/68
