### 2023-05-14 paulbrodersen/netgraph#69 @ortega2247
Minor documentation issues
1. It would be nice to explicitly say that it is necessary to install networkx, igraph, etc, to be able to create the graphs (I wrongly assumed that some of these packages were requirements during the installation)
2. I tried to use EditableGraph on windows and didn't work. It would be nice to have a note that it is not supported on windows (yet?)
