### 2023-05-14 #5372 @ortega2247
Minor documentation issues: https://github.com/paulbrodersen/netgraph/issues/69
