### 2023-05-16 #5372 @ortega2247
The current netgraph github repo has test suite but it is not automated. It would be great to have it automated
