### 2023-05-16 #5372 @ortega2247
The paper could extend a bit and describe the algorithms that are implemented. Additionally, the authors could add details about how "... it handles networks with multiple components gracefully", what they mean by that, and how is better than other network visualization algorithms
