### 2023-05-16 paulbrodersen/netgraph#68 @paulbrodersen
Thanks for raising the issue. Unfortunately, I can't be of help in this case. Netgraph does not contain an `_imaging` module.

A quick google search suggests that `PIL` / `pillow` may be the culprit; to be sure, however, you would have know the complete error trace. If that is indeed the case, I would suggest installing `pillow` separately, or upgrading `pillow` if you have an existing installation. `pillow` is a dependency of matplotlib (IIRC) but netgraph does not use any of the related features so any `pillow` version should work (even none at all) as far as netgraph is concerned.

As I am not author/maintainer of the `_imaging` module, and can't provide further assistance, I will close the issue for now.
