### 2023-05-16 paulbrodersen/netgraph#69 @paulbrodersen
> It would be nice to explicitly say that it is necessary to install networkx, igraph, etc, to be able to create the graphs (I wrongly assumed that some of these packages were requirements during the installation)

Thank you for your suggestion. I will make a corresponding note in the documentation.

> I tried to use EditableGraph on windows and didn't work. It would be nice to have a note that it is not supported on windows (yet?)

I have worked on windows 10 boxes with netgraph, so this would surprise me. Please make sure that you are using an interactive matplotlib backend. Note that jupyter notebooks and google colab notebooks do not support event handling and are hence not interactive. If you are using an IDE like pycharm, make sure to run any interactive code in a shell (Alt+Shift+E in pycharm) as IDEs otherwise also typically don't support event handling. You can find instructions on how to switch backends [here](https://matplotlib.org/stable/users/explain/backends.html).
