### 2023-05-16 paulbrodersen/netgraph#68 @rkurchin
Linking to [JOSS review](https://github.com/openjournals/joss-reviews/issues/5372) and popping in to note: I think if there are installation hiccups on the latest version of Python, we should make sure any gotchas are documented – @ortega2247, can you confirm whether the stacktrace does implicate `pillow` and whether installing NetGraph in a clean conda environment (i.e. without preexisting `pillow` and/or `matplotlib` installations) still causes this issue?
