### 2023-05-16 paulbrodersen/netgraph#69 @rkurchin
Linking to [JOSS review](https://github.com/openjournals/joss-reviews/issues/5372) and dropping into to note that I'm not able to find mention of the requirement of a particular matplotlib backend in the docs, so that would be a useful thing to explicitly include...
