### 2023-05-17 6eefd1b
Improve consistency in the naming of edge layout tests
