### 2023-05-17 b623bef
Make specification of optional requirements consistent between pyproject.toml, setup.py, and optional_dependencies.txt
