### 2023-05-17 643a5e9
Document how to install optional dependencies (issue #69)
