### 2023-05-17 88d12ef
Document how to use the automated test suite
