### 2023-05-17 d7839de
Rename `extras_require[doc]` to `extras_require[docs]`
