+++ pyproject.toml
@@ -42,13 +42,13 @@ tests = [
       "pytest",
       "pytest-mpl",
 ]
-doc = [
-    "sphinx",
-    "sphinx-rtd-theme",
-    "numpydoc",
-    "sphinx-gallery",
-    "Pillow",
-    "networkx",
+docs = [
+     "sphinx",
+     "sphinx-rtd-theme",
+     "numpydoc",
+     "sphinx-gallery",
+     "Pillow",
+     "networkx",
 ]
 
 [tool.setuptools.dynamic]
