### 2023-05-17 paulbrodersen/netgraph#69 @paulbrodersen
Commit [643a5e9](https://github.com/paulbrodersen/netgraph/commit/643a5e95521356e315e2a5ff61232738eb83aba9) (and following) expands the documentation to provide instructions for installing all optional dependencies required to run the examples in the documentation.

Commit [34e200d](https://github.com/paulbrodersen/netgraph/commit/34e200d011ba7e813000a30b67dde6e0b832b195) expands the documentation to provide guidance for troubleshooting issues with matplotlib's event handling. This includes a link to matplotlib's documentation, where all currently available interactive backends are listed. This should hence address also your additional issue, @rkurchin.

I hope this addresses both of your comments. If not, feel free to re-open.
