### 2023-05-17 paulbrodersen/netgraph#68 @paulbrodersen
> I think if there are installation hiccups on the latest version of Python, we should make sure any gotchas are documented

I am sorry but I can't monitor all upstream packages for installation issues on all platforms. I have neither the time nor resources to do so. The `pillow` team is very active. If indeed there is an issue with `pillow` on some windows platforms, I have no way of knowing if and when it is resolved. On my windows installation, `pillow` installs without errors when using conda.
