### 2023-05-17 #5372 @paulbrodersen
Hi @ortega2247, first of all thank you for taking the time to try out my library and review my work.

I have addressed some of your concerns on Netgraph's github issue tracker.
Below, I will summarise these efforts and expand on the issues that weren't addressed there.

---

#### Comments

> Minor issue: paulbrodersen/netgraph#68

As I commented [here](https://github.com/paulbrodersen/netgraph/issues/68#issuecomment-1549792426), this issue seems unrelated to my code base and is hence unfortunately outside the scope of my library and outside of my control.

> Minor documentation issues: paulbrodersen/netgraph#69

I addressed this double issue in two commits. Commit #643a5e9  expands the documentation to provide instructions for installing all optional dependencies required to run the examples in the documentation. Commit #34e200d expands the documentation to provide guidance for troubleshooting issues with matplotlib's event handling and hence Netgraph's interactive graph classes such as `EditableGraph`. This includes a link to matplotlib's documentation, where all currently available interactive backends are listed, and thus hopefully also addresses @rkurchin's additional issue.
Regarding the second issue, I am confident that with a correctly configured matplotlib backend, you will be able to run the `EditableGraph` class (even) with a Windows installation.

> The current netgraph github repo has test suite but it is not automated. It would be great to have it automated

As mentioned in the paper, netgraph uses pytest for automated testing. The test-suite can be executed by switching to the top level directory and running the `pytest` command. This is the standard procedure for any test suite using pytest but I have made another note in the documentation to that effect. The relevant commit can be found [here](88d12ef71e4305cc56e42e2e2776c22623d18463).

> The paper could extend a bit and describe the algorithms that are implemented.

Currently, the node layout and edge routing algorithms and the full API of my implementations are listed [here](https://netgraph.readthedocs.io/en/latest/node_layout.html) and [here](https://netgraph.readthedocs.io/en/latest/edge_layout.html) in the online documentation. In principle, I am happy to add a similar list to the article, and indeed an earlier draft contained such a list. My concern is that the number of implemented node layout algorithms is rapidly growing, and as a result, any explicit list in the journal article would likely be out of date by the time it was read by the reader.

@rkurchin What is JOSS' policy on outgoing links? The example paper does not include any.

If outgoing links are allowed, I would suggest substituting the following sentence in the summary section

```
Netgraph implements numerous node layout algorithms and several edge routing routines."
```

with

```
At the time of writing, Netgraph provides the following node layout algorithms:
- the Fruchterman-Reingold algorithm a.k.a. the "spring" layout,
- the Sugiyama algorithm a.k.a. the "dot" layout for directed, acyclic graphs,
- a radial tree layout for directed, acyclic graphs,
- a circular node layout (with optional edge crossing reduction),
- a bipartite node layout for bipartite graphs (with optional edge crossing reduction),
- a layered node layout for multipartite graphs (with optional edge crossing reduction),
- a shell layout for multipartite graphs (with optional edge crossing reduction),
- a community node layout for modular graphs, and
- a "geometric" node layout for graphs with defined edge lengths but unknown node positions.

Additionally, links or edges between the nodes can be straight, curved (avoiding collisions with other nodes and edges), or bundled.
However, as new layout routines are added regularly to Netgraph, consult the online documentation [here](https://netgraph.readthedocs.io/en/latest/node_layout.html) for an up-to-date list.
```

> Additionally, the authors could add details about how "... it handles networks with multiple components gracefully", what they mean by that, and how is better than other network visualization algorithms

Node layout algorithms are typically only defined for single components. When they are applied to disconnected/multi-component graphs, they often result in numerical instabilities and thus non-interpretable visualisations. For the spring layout a.k.a. the Fruchterman-Reingold algorithm, which is the default node layout algorithm used in all major network analysis python libraries (networkx/igraph/graph-tool), example visualisations and discussions of this problem can be found [here](https://stackoverflow.com/q/53120739/2912349) and [here](https://github.com/networkx/networkx/pull/3488). Unlike other libraries, netgraph wraps the implemented node layout algorithms such that each component is processed individually, thus avoiding the numerical instabilities caused by a naive application of the node layout algorithms to unconnected graphs. Typically, the layouts for the individual components are then arranged w.r.t. each other using a rectangle-packing algorithm implemented in the `rectangle-packer` library. This behaviour is documented [here](https://netgraph.readthedocs.io/en/latest/sphinx_gallery_output/plot_12_multicomponent_graphs.html#sphx-glr-sphinx-gallery-output-plot-12-multicomponent-graphs-py) in the documentation.

Having only briefly outlined the gist of the problem that is being solved by this feature of the library, I hope it is clear that a publishable and hence necessarily more thorough discussion of this issue would not mesh well with the aim to provide "a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience". I would hence like to avoid expanding on this issue in this paper.

---

#### Outstanding items on the checklist

> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

Brief installations instructions are given in the README on github and extensive installation instructions are given in the documentation on ReadTheDocs [here](https://netgraph.readthedocs.io/en/latest/installation.html), both for `pip` and `conda/mamba`. The requirements are documented in various machine readable formats [here](https://github.com/paulbrodersen/netgraph/blob/master/requirements.txt), [here](https://github.com/paulbrodersen/netgraph/blob/master/optional-requirements.txt), [here](https://github.com/paulbrodersen/netgraph/blob/master/pyproject.toml), and [here](https://github.com/paulbrodersen/netgraph/blob/master/setup.py), supporting a wide range of different `pip` versions as well as installation via `conda/mamba` from conda-forge. I would hence appreciate further clarifications why this box is left unchecked.

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

I have addressed this in a comment above.

> Summary: Has a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience been provided?

I am assuming the lack of a checkmark is due to the lack of a complete listing of all available node layout and edge routing algorithms, and are hence addressed by my suggested changes to the text outlined above. Otherwise, further clarifications would be much appreciated.

I hope this addresses all issues raised so far. If you have any further concerns, please let me know at your earliest convenience.
