### 2023-05-17 paulbrodersen/netgraph#68 @rkurchin
Understood. Another thing you could consider would be setting up CI – there are standard Windows runners (and other OSes as well obviously), which would at least let you catch if the "standard" environment/setup trips up, either due to changes in commits you introduce or to upstream packages.
