### 2023-05-17 e34bad9
Small improvements to the matplotlib event handling troubleshooting guide
