### 2023-05-17 paulbrodersen/netgraph#68 @paulbrodersen
@rkurchin Absolutely. Setting up CI is already on my TODO list but I haven't got around to it yet. As there are two major tickets that I would like to address first (major release with support for multigraphs, improved integration tests), it won't happen in time for this review.
