### 2023-05-18 69dcf3a
Prevent RunTimeWarning due to divide-by-zero errors
