### 2023-05-22 e430dbe
Simplify edge path updates
