### 2023-05-22 5852da3
In `get_straight_edge_paths`, compute layout for self-loops separately
