### 2023-05-22 0b6dffc
In `get_arced_edge_paths`, compute self-loop edge paths separately
