### 2023-05-22 d4996a3
In `get_arced_edge_paths`, compute edge paths for all components simultaneously
Handling components individually should only increase computational load.
