### 2023-05-22 ac77fb8
Remove superfluous `_get_selfloop_path`
