### 2023-05-22 094a74a
In `get_curved_edge_paths`, compute self-loop edge paths separately
