+++ netgraph/_edge_layout.py
@@ -134,20 +134,23 @@ def _get_optimal_selfloop_angles(selfloops, selfloop_radius, node_positions, edg
     return dict(zip(selfloops, selfloop_angles))
 
 
-def _get_straight_selfloop_edge_paths(edges, node_positions, selfloop_radius, selfloop_angle):
+def _get_straight_selfloop_edge_paths(edges, node_positions, selfloop_radius, selfloop_angle, total_points=100):
     edge_paths = dict()
     for edge in edges:
-        unit_vector = _get_unit_vector(
-            np.array([np.cos(selfloop_angle[edge]),
-                      np.sin(selfloop_angle[edge])]))
-        center = node_positions[edge[0]] + selfloop_radius[edge] * unit_vector
-        # Note: we add pi to the start angle as the start angle lies opposite
-        # to the direction in which the self-loop extends.
-        edge_paths[edge] = _get_n_points_on_a_circle(
-            center, selfloop_radius[edge], 100+1,
-            _get_angle(*unit_vector) + np.pi,
-        )[1:]
+        edge_paths[edge] = _get_selfloop_path(
+            node_positions[edge[0]], selfloop_radius[edge], selfloop_angle[edge], total_points)
     return edge_paths
+
+
+def _get_selfloop_path(source_position, radius, angle, total_points):
+    unit_vector = _get_unit_vector(np.array([np.cos(angle), np.sin(angle)]))
+    center = source_position + radius * unit_vector
+    # Note: we add pi to the start angle as the start angle lies opposite
+    # to the direction in which the self-loop extends.
+    return _get_n_points_on_a_circle(
+        center, radius, total_points+1,
+        _get_angle(*unit_vector) + np.pi,
+    )[1:]
 
 
 def get_curved_edge_paths(edges, node_positions,
@@ -203,10 +206,39 @@ def get_curved_edge_paths(edges, node_positions,
 
     """
 
-    edge_to_control_points = _initialize_control_points(edges, node_positions, scale)
+    nonloops = [(source, target) for (source, target) in edges if source != target]
+    selfloops = [(source, target) for (source, target) in edges if source == target]
+    selfloop_radius = _normalize_numeric_argument(selfloop_radius, selfloops, 'selfloop_radius')
 
-    control_point_positions = _initialize_control_point_positions(
-        edge_to_control_points, node_positions, selfloop_radius, selfloop_angle)
+    nonloop_edge_paths = _get_curved_nonloop_edge_paths(
+        nonloops, node_positions, origin, scale, k, initial_temperature,
+        total_iterations, node_size, bundle_parallel_edges)
+
+    if selfloop_angle is not None: # can be zero!
+        selfloop_angle = _normalize_numeric_argument(selfloop_angle, selfloops, 'angle', allow_none=True)
+    else:
+        selfloop_angle = _get_optimal_selfloop_angles(
+            selfloops, selfloop_radius, node_positions, nonloop_edge_paths)
+
+    selfloop_edge_paths = _get_curved_selfloop_edge_paths(
+        selfloops, node_positions, selfloop_radius, selfloop_angle,
+        origin, scale, k, initial_temperature,
+        total_iterations, node_size, nonloop_edge_paths)
+
+    edge_paths = dict()
+    edge_paths.update(nonloop_edge_paths)
+    edge_paths.update(selfloop_edge_paths)
+    return edge_paths
+
+
+def _get_curved_nonloop_edge_paths(edges, node_positions, origin, scale,
+                                   k, initial_temperature, total_iterations,
+                                   node_size, bundle_parallel_edges):
+
+    edge_to_control_points = _initialize_nonloop_control_points(edges, node_positions, scale)
+
[…]
