+++ publication/paper.md
@@ -29,7 +29,6 @@ The empirical study and scholarly analysis of networks has increased manyfold in
 
 Netgraph is a python library that aims to complement the existing network analysis libraries with publication quality visualisations within the python ecosystem. To facilitate a seamless integration, netgraph supports a variety of input formats, including networkx, igraph, and graph-tool Graph objects. At the time of writing, Netgraph provides the following node layout algorithms:
 
-Netgraph is licensed under the General Public License version 3 (GPLv3). The repository is hosted on github, and distributed via PyPI and conda-forge. It includes an extensive automated test suite that can be executed using pytest. The comprehensive documentation -- including a complete API reference as well as numerous examples and tutorials -- is hosted on ReadTheDocs.
 - the Fruchterman-Reingold algorithm a.k.a. the "spring" layout,
 - the Sugiyama algorithm a.k.a. the "dot" layout for directed, acyclic graphs,
 - a radial tree layout for directed, acyclic graphs,
@@ -45,6 +44,7 @@ However, new layout routines are added regularly to Netgraph; for an up-to-date 
 
 Uniquely among python alternatives, Netgraph handles networks with multiple components gracefully (which otherwise break most node layout routines), and it post-processes the output of the node layout and edge routing algorithms with several heuristics to increase the interpretability of the visualisation (reduction of overlaps between nodes, edges, and labels; edge crossing minimisation and edge unbundling where applicable). The highly customisable plots are created using matplotlib [@Hunter:2007], a popular python plotting library, and the resulting matplotlib objects are exposed in an easily queryable format such that they can be further manipulated and/or animated using standard matplotlib syntax. The visualisations can also be altered interactively: nodes and edges can be added on-the-fly through hotkeys, positioned using the mouse, and labelled through standard text-entry.
 
+Netgraph is licensed under the General Public License version 3 (GPLv3). The repository is hosted on [github](https://github.com/paulbrodersen/netgraph), and distributed via PyPI and conda-forge. It includes an extensive automated test suite that can be executed using pytest. The comprehensive documentation -- including a complete API reference as well as numerous examples and tutorials -- is hosted on [ReadTheDocs](https://netgraph.readthedocs.io).
 
 # Figures
 
