### 2023-05-30 e4729a7
fixup! In `get_arced_edge_paths`, compute self-loop edge paths separately
