### 2023-05-30 fc90e3b
Only compute selfloop properties when there are any
