### 2023-05-31 d833754
Bugfix: ensure that bundled parallel edges have the same path
