### 2023-05-31 a98da08
Support non-standard repulsion and attraction models in FR layout
This allows us to remove the code duplication in _edge_layout.
