+++ netgraph/_edge_layout.py
@@ -365,7 +365,7 @@ def _optimize_control_point_positions(
 
         if bundle_parallel_edges:
             # Edge control points are repulsed by nodes but not by other edge control points.
-            expanded_node_positions = _get_fruchterman_reingold_layout(
+            expanded_node_positions = get_fruchterman_reingold_layout.__wrapped__(
                 expanded_edges,
                 node_positions      = expanded_node_positions,
                 scale               = scale,
@@ -375,6 +375,7 @@ def _optimize_control_point_positions(
                 total_iterations    = total_iterations,
                 node_size           = node_size,
                 fixed_nodes         = nodes,
+                get_repulsion       = _get_fr_repulsion_variant,
             )
         else:
             # Edge control points are repulsed by other edge control points.
@@ -407,167 +408,7 @@ def _expand_edges(edge_to_control_points):
     return expanded_edges
 
 
-# This is a (slightly simplified) copy of the function defined in _node_layout.
-# This allows us to redefine the internally called function _get_fr_repulsion.
-# TODO: find a way to avoid code repetition.
-# NOTE: Monkey patching did not work as intended (commit 435f187f99b8ff43d1d573c5e9302ea92cfa7eb2).
-def _get_fruchterman_reingold_layout(edges,
-                                     edge_weights        = None,
-                                     k                   = None,
-                                     scale               = None,
-                                     origin              = None,
-                                     initial_temperature = 1.,
-                                     total_iterations    = 50,
-                                     node_size           = 0,
-                                     node_positions      = None,
-                                     fixed_nodes         = None,
-                                     *args, **kwargs):
-    """'Spring' or Fruchterman-Reingold node layout.
-
-    Uses the Fruchterman-Reingold algorithm [Fruchterman1991]_ to compute node positions.
-    This algorithm simulates the graph as a physical system, in which nodes repell each other.
-    For connected nodes, this repulsion is counteracted by an attractive force exerted by the edges, which are simulated as springs.
-    The resulting layout is hence often referred to as a 'spring' layout.
-
-    Parameters
-    ----------
-    edges : list
-        The edges of the graph, with each edge being represented by a (source node ID, target node ID) tuple.
-    edge_weights : dict
-        Mapping of edges to edge weights.
-    k : float or None, default None
-        Expected mean edge length. If None, initialized to the sqrt(area / total nodes).
-    origin : tuple or None, default None
-        The (float x, float y) coordinates corresponding to the lower left hand corner of the bounding box specifying the extent of the canvas.
-        If None is given, the origin is placed at (0, 0).
-    scale : tuple or None, default None
-        The (float x, float y) dimensions representing the width and height of the bounding box specifying the extent of the canvas.
-        If None is given, the scale is set to (1, 1).
-    total_iterations : int, default 50
-        Number of iterations.
-    initial_temperature: float, default 1.
-        Temperature controls the maximum node displacement on each iteration.
-        Temperature is decreased on each iteration to eventually force the algorithm into a particular solution.
-        The size of the initial temperature determines how quickly that happens.
-        Values should be much smaller than the values of `scale`.
-    node_size : scalar or dict, default 0.
-        Size (radius) of nodes.
-        Providing the correct node size minimises the overlap of nodes in the graph,
[…]
