### 2023-06-02 1b8906f
Implement StraightEdgeLayout class
