+++ netgraph/_main.py
@@ -46,10 +46,7 @@ from ._node_layout import (
 )
 from ._edge_layout import (
     _shift_edge,
-    get_straight_edge_paths,
-    get_curved_edge_paths,
-    get_arced_edge_paths,
-    get_bundled_edge_paths,
+    StraightEdgeLayout,
 )
 
 from ._artists import (
@@ -327,7 +324,7 @@ class BaseGraph(object):
                         node_shape, node_size, node_edge_width,
                         node_color, node_edge_color, node_alpha, node_zorder)
 
-        self.edge_paths, self.edge_layout, self.edge_layout_kwargs = self._initialize_edge_layout(
+        self.edge_paths, self.edge_layout = self._initialize_edge_layout(
             edge_layout, edge_layout_kwargs, origin, scale, edge_width, self.node_artists)
         self.edge_artists = dict()
         self.draw_edges(self.edge_paths, edge_width, edge_color, edge_alpha,
@@ -472,22 +469,17 @@ class BaseGraph(object):
             pass
 
         if isinstance(edge_layout, str):
-            edge_paths = self._get_edge_paths(self.edges, self.node_positions,
-                                              edge_layout, edge_layout_kwargs)
+            edge_layout = StraightEdgeLayout(self.edges, self.node_positions, **edge_layout_kwargs)
+            edge_paths = edge_layout.compute()
         elif isinstance(edge_layout, dict):
             _check_completeness(edge_layout, self.edges, 'edge_layout')
             edge_paths = edge_layout
-
-            # determine a sensible edge_layout in case node positions change
-            path_lengths = np.array([len(path) for path in edge_paths.values()])
-            if np.any(path_lengths) > 2:
-                edge_layout = 'curved'
-            else:
-                edge_layout = 'straight'
+            edge_layout = StraightEdgeLayout(self.edges, self.node_positions, **edge_layout_kwargs)
+            edge_layout.edge_paths.update(edge_paths)
         else:
             raise TypeError("Variable `edge_layout` either a string or a dict mapping edges to edge paths.")
 
-        return edge_paths, edge_layout, edge_layout_kwargs
+        return edge_paths, edge_layout
 
 
     def _initialize_axis(self, ax):
@@ -583,51 +575,6 @@ class BaseGraph(object):
             self.node_artists[node].xy = self.node_positions[node]
 
 
-    def _get_edge_paths(self, edges, node_positions, edge_layout, edge_layout_kwargs):
-        """Compute the edge routing.
-
-        Parameters
-        ----------
-        edges : list
-            The edges of the graph, with each edge being represented by a (source node ID, target node ID) tuple.
-        node_positions : dict
-            Mapping of nodes to (x, y) positions
-        edge_layout : 'straight', 'curved' or 'bundled' (default 'straight')
-            If 'straight', draw edges as straight lines.
-            If 'curved', draw edges as curved splines. The spline control points are optimised to avoid other nodes and edges.
-            If 'bundled', draw edges as edge bundles.
-        edge_layout_kwargs : dict
-            Keyword arguments passed to edge layout functions.
-            See the documentation of the following functions for a full list of available options:
-            - get_straight_edge_paths
-            - get_curved_edge_paths
-            - get_bundled_edge_paths
-
-        Returns
-        -------
-        edge_paths : dict
-            Mapping of edges to arrays of (x, y) tuples, the edge path coordinates.
-
-        """
-
-        if edge_layout == 'straight':
-            edge_paths = get_straight_edge_paths(
-                edges, node_positions, **edge_layout_kwargs)
-        elif edge_layout == 'curved':
[…]
