### 2023-06-07 2b5b014
Implement CurvedEdgeLayout class
