### 2023-06-09 70a1e65
Compute updates to curved edge layout using straight edge layout routines
