### 2023-06-09 339b0ef
Implement BundledEdgeLayout class
