### 2023-06-13 f049f7c
Reduce time required to run edge bundling example
