### 2023-06-13 b91068b
Correct order of decorators
The profile decorator has to be the inner decorator; otherwise,
_handle_multiple_components is profiled, not get_bundled_edges
