### 2023-06-13 517cd5d
Add progress bar to `get_bundled_edge_paths`
