### 2023-06-13 62cde20
Determine components without recursion (issue #67)
Determining the connected components in very large graphs with
recursive implementations of DFS or BFS can result in exceeding the
recursion limit in python.
