### 2023-06-13 0b6d4ac
Raise a warning if the plot is guaranteed to result in a hairball
