### 2023-06-13 4ec3095
Specify correct minimum python version
f-strings are only supported since python 3.6
