### 2023-06-19 #5372 @idoby
> Hi @idoby! Thanks for reviewing my work. If you have any questions or concerns, do let me know -- either here, or in netgraph's issue tracker. Cheers! Paul

Thanks for publishing your package!
I'll only get to doing this review later this week, I'm afraid, but this package does look interesting. :)
