+++ docs/source/sphinx_gallery_output/sg_execution_times.rst
@@ -5,10 +5,10 @@
 
 Computation times
 =================
-**00:05.051** total execution time for **sphinx_gallery_output** files:
+**03:29.683** total execution time for **sphinx_gallery_output** files:
 
 +-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_sphinx_gallery_output_plot_20_geometric_layout.py` (``plot_20_geometric_layout.py``)                                     | 00:05.051 | 0.0 MB |
+| :ref:`sphx_glr_sphinx_gallery_output_plot_10_community_layout.py` (``plot_10_community_layout.py``)                                     | 03:29.683 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_sphinx_gallery_output_plot_01_default.py` (``plot_01_default.py``)                                                       | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
@@ -28,8 +28,6 @@ Computation times
 +-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_sphinx_gallery_output_plot_09_circular_layout.py` (``plot_09_circular_layout.py``)                                       | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
-| :ref:`sphx_glr_sphinx_gallery_output_plot_10_community_layout.py` (``plot_10_community_layout.py``)                                     | 00:00.000 | 0.0 MB |
-+-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_sphinx_gallery_output_plot_11_curved_edge_layout.py` (``plot_11_curved_edge_layout.py``)                                 | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_sphinx_gallery_output_plot_12_multicomponent_graphs.py` (``plot_12_multicomponent_graphs.py``)                           | 00:00.000 | 0.0 MB |
@@ -47,4 +45,10 @@ Computation times
 | :ref:`sphx_glr_sphinx_gallery_output_plot_18_legends.py` (``plot_18_legends.py``)                                                       | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
 | :ref:`sphx_glr_sphinx_gallery_output_plot_19_hyperlinks.py` (``plot_19_hyperlinks.py``)                                                 | 00:00.000 | 0.0 MB |
++-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_sphinx_gallery_output_plot_20_geometric_layout.py` (``plot_20_geometric_layout.py``)                                     | 00:00.000 | 0.0 MB |
++-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_sphinx_gallery_output_plot_21_flowchart.py` (``plot_21_flowchart.py``)                                                   | 00:00.000 | 0.0 MB |
++-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+
+| :ref:`sphx_glr_sphinx_gallery_output_plot_22_large_graphs.py` (``plot_22_large_graphs.py``)                                             | 00:00.000 | 0.0 MB |
 +-----------------------------------------------------------------------------------------------------------------------------------------+-----------+--------+

[…]
