### 2023-06-19 baebbb6
Bug fix: add missing `verbose` argument to BundledEdgeLayout class
