### 2023-06-19 60c6bef
Bug fix: work around memory error raised by rpack (issue #67)
