+++ netgraph/_node_layout.py
@@ -160,14 +160,14 @@ def _get_packed_component_bboxes(components, origin, scale, power=0.8, pad_by=0.
 
     # Add a padding between boxes, such that nodes cannot end up touching in the final layout.
     # We choose a padding proportional to the dimensions of the largest box.
-    maximum_dimensions = np.max(relative_dimensions, axis=0)
-    pad_x, pad_y = pad_by * maximum_dimensions
+    maximum_width, maximum_height = np.max(relative_dimensions, axis=0)
+    pad_x, pad_y = pad_by * maximum_width, pad_by * maximum_height
     padded_dimensions = [(width + pad_x, height + pad_y) for (width, height) in relative_dimensions]
 
     # rpack only works on integers, hence multiply by some large scalar to retain some precision;
     # NB: for some strange reason, rpack's running time is sensitive to the size of the boxes, so don't make the scalar too large
     # TODO find alternative to rpack
-    scalar = 20
+    scalar = 1000 / ((1 + pad_by) * max(maximum_width, maximum_height))
     integer_dimensions = [(int(scalar*width), int(scalar*height)) for width, height in padded_dimensions]
     origins = pack(integer_dimensions) # NB: rpack claims to return upper-left corners, when it actually returns lower-left corners
 
