### 2023-06-27 #5372 @idoby
@paulbrodersen I took a look at the paper and the code, and I'm liking the package very much! It seems useful and easy to use, especially the interactive and editable graph features.

However, I found the quality of the paper itself can be improved. I'm not sure this warrants opening an issue in the repo, so I'll list my main points here:

1. The paper is quite short and does not fully encompass the capabilities of the software. JOSS papers do tend to run short, but in this case, I think the interactive and editable graph features should be presented, perhaps along with the diagrams shown in the GitHub readme.
2. An example of the basic usage of the software could also be nice, so that the reader could gauge how easy it is to achieve one of the examples shown in the figure, for example. In addition, I feel that the integration with matplotlib is a key feature and should be showcased more.
3. Perhaps dedicate a paragraph or two to the design considerations undertaken by the author of the software, if any interesting design decisions or tradeoffs were made.
4. Minor quality improvements, e.g., "python" should be capitalized ("Python"), GitHub as well, etc.

Overall I think this is quality work and would recommend acceptance after addressing some or all of the above concerns. I feel that the above would improve the quality of the paper and increase reader interest while still maintaining a short and readable paper.
