### 2023-06-27 8fb1ede
Correct the capitalisation of Python, GitHub, NetworkX, and Matplotlib
