### 2023-06-27 #5372 @paulbrodersen
> Minor quality improvements, e.g., "python" should be capitalized ("Python"), GitHub as well, etc.

Corrected the capitalization for Python, GitHub, NetworkX, and Matplotlib in [this commit](https://github.com/paulbrodersen/netgraph/commit/8fb1ede6e0c1bdf6aafacc6460e51407912009dc).
