+++ publication/paper.md
@@ -48,9 +48,11 @@ Netgraph is licensed under the General Public License version 3 (GPLv3). The rep
 
 # Figures
 
-![Netgraph's key distinguishing features](gallery_portrait.png){width=90%}
+![Showcase of Netgraph's key distinguishing features](gallery_portrait.png){width=90%}
 
-# Basic Example Usage
+<div style="page-break-after: always;"></div>
+
+# A Basic Example
 
 The following script shows a minimum, working example. The graph structure is defined by an edge list, and the visualisation is created using (mostly) default parameters.
 
@@ -73,15 +75,11 @@ Graph(
 plt.show()
 ```
 
-![Basic example output](basic_example.png){width=66%}
+![Basic example output](basic_example.png){width=50%}
 
-# Advanced Example Usage
+# Interoperability & Customizability
 
-The next example demonstrates more advanced features of the library.
-
-Netgraph can be easily integrated into existing network analysis workflows as it accepts a variety of graph structures. The example below uses a NetworkX `Graph` object, but igraph and graph-tool objects are also valid inputs, as are plain edge lists and full-rank adjacency matrices.
-
-The output visualizations are created using Matplotlib and can hence form subplots in larger Matplotlib figures.
+Netgraph can be easily integrated into existing network analysis workflows as it accepts a variety of graph structures. The example below uses a NetworkX `Graph` object, but igraph and graph-tool objects are also valid inputs, as are plain edge lists and full-rank adjacency matrices. The output visualizations are created using Matplotlib and can hence form subplots in larger Matplotlib figures.
 
 Each visualization can be customized in various ways. Most parameters can be set using a scalar or string. In this case, the value is applied to all relevant artists. To style each artist differently, supply a dictionary instead. Furthermore, node and edge artists are derived from `matplotlib.patches.PathPatch`; node and edge labels are `matplotlib.text.Text` instances. Hence all node artists, edge artists, and labels can be manipulated using standard matplotlib syntax after the initial draw.
 
@@ -129,7 +127,7 @@ for node in balanced_tree:
 fig.canvas.draw()
 ```
 
-![Advanced example output](advanced_example.png){width=66%}
+![Advanced example output](advanced_example.png){width=50%}
 
 # Acknowledgements
 
