### 2023-06-27 7fd3797
Show output figures before the code that generated them
