### 2023-06-27 #5372 @paulbrodersen
> An example of the basic usage of the software could also be nice, so that the reader could gauge how easy it is to achieve one of the examples shown in the figure, for example. In addition, I feel that the integration with matplotlib is a key feature and should be showcased more.

I have added two sections. "A Basic Example" demonstrates a minimum working example to create a figure with Netgraph. The second section, "Interoperability & Customizability", demonstrates more advanced customizations and the integration with matplotlib. Both examples are (loosely) inspired by examples in the first figure. The basic example is a slightly simplified version, whereas the advanced example is more sophisticated version of its counterpart in the first figure.

As these changes span multiple commits, it is probably easier to review them in the PDF.
