### 2023-06-27 1eb8563
Discuss merits of having a single reference frame
