+++ publication/paper.md
@@ -131,6 +131,14 @@ for node in balanced_tree:
 fig.canvas.draw()
 ```
 
+# Key Design Decisions
+
+The creation of Netgraph was motivated by the desire to make high quality, easily customizable, and reproducible network vizualisations, whilst maintaining an extensible code base. To that end, a key design decision was to have a single reference frame for all node artist and edge artist attributes that determine their extent (e.g. in the case of a circular node artist, its position and its radius).
+
+Good data visualizations are both accurate and legible. The legibility of a visualization is influenced predominantly by the size of the plot elements, and occlusions between them. However, there is often a tension between these two requirements, as larger plot elements are more visible but also more likely to cause overlaps with other plot elements. Most data visualization tools focus on accuracy and visibility. To that end, they operate in two reference frames: a data-derived reference frame and a display-derived reference frame. For example, in a standard line-plot, the data-derived reference frame determines the x and y values of the line. The thickness of the line, however, scales with the size of the display, and its width (measured in pixels) remains constant across different figure sizes and aspect ratios. Having two reference frames ensures that the line (1) is an accurate representation of the data, and (2) is visible and discernible independent of figure dimensions. The trade-off of this setup is that (1) the precise extents of plot elements can only be computed after the figure is initialized, and (2) occlusions are not managed and hence common, for example, if multiple lines are plotted in the same figure. Nevertheless, most network visualization tools follow this standard. For example, NetworkX specifies node positions and edge paths in data coordinates, but uses display units for node sizes and edge widths.
+
+However, network visualizations differ from other data visualizations in two aspects: (1) the precise positions of nodes and the precise paths of edges often carry no inherent meaning, and (2) most figures contain a multitude of node and edge artists instead of just the few lines typically present in a line-plot. As a consequence, a common goal of most algorithms for node layout, edge routing, and label placement is to minimize occlusions between different plot elements, as they reduce the ease with which a visualization is interpreted. To that end, precise knowledge of the extents of all plot elements is paramount, motivating the use of a single reference frame. In Netgraph, this reference frame derives from the data. Specifically, node positions and edge paths are specified in data units, and node sizes and edge widths are specified in 1/100s of data units (as this makes the node sizes and edge widths more comparable to typical values in NetworkX, igraph, and graph-tool). This decouples layout computations from rendering the figure, simplifies computing the extent of the different plot elements, facilitates the reduction of overlaps, and makes it possible to create pixel-perfect reproductions independent of display parameters.
+
 # Acknowledgements
 
 We thank GitHub users adleris, Allan L. R. Hansen, chenghuzi, Hamed Mohammadpour, and Pablo for contributing various bug fixes.
