### 2023-06-27 #5372 @paulbrodersen
> Perhaps dedicate a paragraph or two to the design considerations undertaken by the author of the software, if any interesting design decisions or tradeoffs were made.

I have written a draft section (text below) discussing the original raison d'etre for Netgraph, and the core design decision that derived from it. The text is still a little bit clunky, and I will revisit it tomorrow before committing, but before I spend even more time on it, **I would appreciate some feedback if this is what you had in mind, @idoby?**

---

The creation of Netgraph was motivated by the desire to make high quality, easily customizable, and reproducible network visualizations, whilst maintaining an extensible code base. To that end, a core design decision was to have a single, data-derived reference frame for all node artist and edge artist attributes that determine their extent (e.g. in the case of a circular node artist, its position and its radius). Precise knowledge of the extents of all plot elements is paramount, as a common goal of most algorithms for node layout, edge routing, and label placement is the reduction or outright avoidance of overlaps between different plot elements, as such overlaps reduce the ease with which a visualization is interpreted.

However, in existing Python libraries, different attributes of a plot element often use fundamentally different reference frames. For example, NetworkX specifies node positions and edge paths in data coordinates, but uses display units for node sizes and edge widths. The transformation from data coordinates to display coordinates is only known once the figure is initialized. In the code base, this results in a tight coupling between computing the extent of plot elements and drawing them. Also, it complicates pre-computing the exact extents of node and edge artists, and reducing or avoiding overlaps between them. Finally, visualizations often cannot be exactly reproduced with different figure renderers, or between different computing environments, if their display resolution or display size differ.

Netgraph amends these issues by having a single reference frame that derives from the data. Specifically, node positions and edge paths are specified in data units, and node sizes and edge widths are specified in 1/100s of data units (as this makes the node sizes and edge widths more comparable to typical node sizes in NetworkX, igraph, and graph-tool). This simplifies computing the extent of the different plot elements, facilitates the reduction of overlaps, decouples layout computations from rendering the figure, and makes it possible to create pixel-perfect reproductions independent of display parameters.
