### 2023-06-27 #5372 @paulbrodersen
>  I think the interactive and editable graph features should be presented

The current version of the manuscript already includes the following -- though admittedly brief -- description:

> The visualisations can also be altered interactively: nodes and edges can be added on-the-fly through hotkeys, positioned using the mouse, and labelled through standard text-entry.

In the documentation, the GIFs do a reasonable job of succinctly demonstrating these capabilities ([link](https://netgraph.readthedocs.io/en/latest/interactivity.html)). However, I am struggling with translating these to a description for a static journal article, without explaining more or less the full API, which would run afoul JOSS's guidelines:

 > [S]oftware documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.

**Do you have any suggestions, or could you explain more thoroughly, how I could expand on the current description, @idoby?**
