### 2023-06-27 #5372 @idoby
> > Perhaps dedicate a paragraph or two to the design considerations undertaken by the author of the software, if any interesting design decisions or tradeoffs were made.
>
> I have written a draft section (text below) discussing the original raison d'etre for Netgraph, and the core design decision that derived from it. The text is still a little bit clunky, and I will revisit it tomorrow before committing, but before I spend even more time on it, **I would appreciate some feedback if this is what you had in mind, @idoby?**

Indeed, this is an interesting bit of motivation/design for the library, I do recommend you include it in the paper.

> In the documentation, the GIFs do a reasonable job of succinctly demonstrating these capabilities ([link](https://netgraph.readthedocs.io/en/latest/interactivity.html)). However, I am struggling with translating these to a description for a static journal article, without explaining more or less the full API, which would run afoul JOSS's guidelines:
>
> > [S]oftware documentation such as API (Application Programming Interface) functionality should not be in the paper and instead should be outlined in the software documentation.
>
> **Do you have any suggestions, or could you explain more thoroughly, how I could expand on the current description, @idoby?**

I thought PDFs were able to embed GIFs. If that's difficult, then I do not suggest going out of your way to include the diagram (nor a description of the API, of course, that is what the docs are for). I would suggest pointing the reader to the readme or docs for more details about the interactive/editable graph feature.

In summary, I suggest adding just the bit above about the design motivation, after editing it to your liking. Otherwise the paper looks good to me. 👍
