+++ publication/paper.md
@@ -133,7 +133,7 @@ fig.canvas.draw()
 
 # Key Design Decisions
 
-The creation of Netgraph was motivated by the desire to make high quality, easily customizable, and reproducible network vizualisations, whilst maintaining an extensible code base. To that end, a key design decision was to have a single reference frame for all node artist and edge artist attributes that determine their extent (e.g. in the case of a circular node artist, its position and its radius).
+The creation of Netgraph was motivated by the desire to make high quality, easily customizable, and reproducible network visualizations, whilst maintaining an extensible code base. To that end, a key design decision was to have a single reference frame for all node artist and edge artist attributes that determine their extent (e.g. in the case of a circular node artist, its position and its radius).
 
 Good data visualizations are both accurate and legible. The legibility of a visualization is influenced predominantly by the size of the plot elements, and occlusions between them. However, there is often a tension between these two requirements, as larger plot elements are more visible but also more likely to cause overlaps with other plot elements. Most data visualization tools focus on accuracy and visibility. To that end, they operate in two reference frames: a data-derived reference frame and a display-derived reference frame. For example, in a standard line-plot, the data-derived reference frame determines the x and y values of the line. The thickness of the line, however, scales with the size of the display, and its width (measured in pixels) remains constant across different figure sizes and aspect ratios. Having two reference frames ensures that the line (1) is an accurate representation of the data, and (2) is visible and discernible independent of figure dimensions. The trade-off of this setup is that (1) the precise extents of plot elements can only be computed after the figure is initialized, and (2) occlusions are not managed and hence common, for example, if multiple lines are plotted in the same figure. Nevertheless, most network visualization tools follow this standard. For example, NetworkX specifies node positions and edge paths in data coordinates, but uses display units for node sizes and edge widths.
 
