### 2023-06-28 #5372 @paulbrodersen
> I thought PDFs were able to embed GIFs.

It is not a widely supported feature. Apparently, even just within Adobe Acrobat products, some do and some don't.

> I would suggest pointing the reader to the readme or docs for more details about the interactive/editable graph feature.

I have added a reference and link to the online tutorial on Netgraph's interactive features in [this commit](https://github.com/paulbrodersen/netgraph/commit/7c368280ea19e8defbb08c903eb779a322392290).
