+++ netgraph/_main.py
@@ -180,9 +180,9 @@ class BaseGraph(object):
         If the type is dict, maps each edge to an individual transparency.
     edge_zorder : int or dict, default 1
         Order in which to plot the edges.
-        If the type is an int, all nodes have the same zorder.
-        If the type is dict, maps each node to an individual zorder.
-        If None, the edges will be plotted in the order they appear in 'adjacency'.
+        If the type is an int, all edges have the same zorder.
+        If the type is dict, maps each edge to an individual zorder.
+        If None, the edges will be plotted in the order they appear in the 'graph' argument.
         Hint: graphs typically appear more visually pleasing if darker edges are plotted on top of lighter edges.
     arrows : bool, default False
         If True, draw edges with arrow heads.
@@ -1062,9 +1062,9 @@ class Graph(BaseGraph):
         If the type is dict, maps each edge to an individual transparency.
     edge_zorder : int or dict, default 1
         Order in which to plot the edges.
-        If the type is an int, all nodes have the same zorder.
-        If the type is dict, maps each node to an individual zorder.
-        If None, the edges will be plotted in the order they appear in 'adjacency'.
+        If the type is an int, all edges have the same zorder.
+        If the type is dict, maps each edge to an individual zorder.
+        If None, the edges will be plotted in the order they appear in the 'graph' argument.
         Hint: graphs typically appear more visually pleasing if darker edges are plotted on top of lighter edges.
     arrows : bool, default False
         If True, draw edges with arrow heads.
@@ -2202,9 +2202,9 @@ class InteractiveGraph(DraggableGraphWithGridMode, EmphasizeOnHoverGraph, Annota
         If the type is dict, maps each edge to an individual transparency.
     edge_zorder : int or dict, default 1
         Order in which to plot the edges.
-        If the type is an int, all nodes have the same zorder.
-        If the type is dict, maps each node to an individual zorder.
-        If None, the edges will be plotted in the order they appear in 'adjacency'.
+        If the type is an int, all edges have the same zorder.
+        If the type is dict, maps each edge to an individual zorder.
+        If None, the edges will be plotted in the order they appear in the 'graph' argument.
         Hint: graphs typically appear more visually pleasing if darker edges are plotted on top of lighter edges.
     arrows : bool, default False
         If True, draw edges with arrow heads.
