### 2023-07-05 71179f6
Rename `shorten_line_by` to `shorten_spline_by`
