### 2023-07-05 23fa070
Fix typo in documentation
