### 2023-07-05 bc4910d
Make input validation in get_fruchterman_reingold_layout optional
Input validation gets in the way when computing curved edge paths
