### 2023-07-05 c96df19
Prepare MultiGraph implementation
