+++ netgraph/_main.py
@@ -294,7 +294,7 @@ class BaseGraph(object):
         self.ax = self._initialize_axis(ax)
         self.fig = self.ax.get_figure()
 
-        self.edges = _parse_edge_list(edges)
+        self.edges = self._initialize_edges(edges)
         self.nodes = self._initialize_nodes(nodes)
 
         self._raise_warning_if_graph_too_large()
@@ -317,22 +317,23 @@ class BaseGraph(object):
         node_edge_width = _rescale_dict_values(node_edge_width, BASE_SCALE)
         edge_width      = _rescale_dict_values(edge_width, BASE_SCALE)
 
-        self.node_size = node_size
-
-        # Initialise node and edge layouts and draw elements.
+        self.node_size = node_size # TODO: replace with node_artist.radius
+        self.edge_width = edge_width
         self.origin = origin
         self.scale = scale
+
+        # Initialise node and edge layouts and draw elements.
         self.node_positions = self._initialize_node_layout(
-            node_layout, node_layout_kwargs, origin, scale, node_size)
+            node_layout, node_layout_kwargs)
         self.node_artists = dict()
         self.draw_nodes(self.nodes, self.node_positions,
-                        node_shape, node_size, node_edge_width,
+                        node_shape, self.node_size, node_edge_width,
                         node_color, node_edge_color, node_alpha, node_zorder)
 
         self.edge_paths, self.edge_layout = self._initialize_edge_layout(
-            edge_layout, edge_layout_kwargs, origin, scale, edge_width, self.node_artists)
+            edge_layout, edge_layout_kwargs)
         self.edge_artists = dict()
-        self.draw_edges(self.edge_paths, edge_width, edge_color, edge_alpha,
+        self.draw_edges(self.edges, self.edge_paths, edge_width, edge_color, edge_alpha,
                         edge_zorder, arrows, self.node_artists)
 
         # This function needs to be called before any font sizes are adjusted,
@@ -367,6 +368,10 @@ class BaseGraph(object):
                                   self.edge_label_rotate, self.edge_label_fontdict)
 
 
+    def _initialize_edges(self, edges):
+        return _parse_edge_list(edges)
+
+
     def _initialize_nodes(self, nodes):
         nodes_in_edges = _get_unique_nodes(self.edges)
         if nodes is None:
@@ -395,103 +400,109 @@ class BaseGraph(object):
             warnings.warn(msg)
 
 
-    def _initialize_node_layout(self, node_layout, node_layout_kwargs, origin, scale, node_size):
+    def _initialize_node_layout(self, node_layout, node_layout_kwargs):
         if node_layout_kwargs is None:
             node_layout_kwargs = dict()
 
         if isinstance(node_layout, str):
             if (node_layout == 'spring') or (node_layout == 'dot') or (node_layout == 'radial'):
-                node_layout_kwargs.setdefault('node_size', node_size)
-            return self._get_node_positions(node_layout, node_layout_kwargs, origin, scale)
+                node_layout_kwargs.setdefault('node_size', self.node_size)
+            return self._get_node_positions(
+                node_layout, node_layout_kwargs, self.edges,
+            )
 
         elif isinstance(node_layout, dict):
             _check_completeness(set(node_layout), set(self.nodes), 'node_layout')
+            # TODO check that nodes are within bounding box set by origin and scale
             return node_layout
 
 
-    def _get_node_positions(self, node_layout, node_layout_kwargs, origin, scale):
+    def _get_node_positions(self, node_layout, node_layout_kwargs, edges):
         if len(self.nodes) == 1:
-            return {self.nodes[0]: np.array([origin[0] + 0.5 * scale[0], origin[1] + 0.5 * scale[1]])}
[…]
