### 2023-07-05 1b1e268
Implement `BaseMultiGraph` class
