+++ tests/test_artists.py
@@ -228,7 +228,8 @@ def plot_edge(x, y, shape='full'):
                        width       = 0.05,
                        head_width  = 0.1,
                        head_length = 0.15,
-                       offset      = 0.1,
+                       head_offset = 0.1,
+                       tail_offset = 0.1,
                        facecolor   = 'red',
                        edgecolor   = 'black',
                        linewidth   = 0.005,
@@ -253,7 +254,8 @@ def test_update_width():
                            width       = 0.05,
                            head_width  = 0.1,
                            head_length = 0.15,
-                           offset      = 0.1,
+                           head_offset = 0.1,
+                           tail_offset = 0.1,
                            facecolor   = 'red',
                            edgecolor   = 'black',
                            linewidth   = 0.005,
