### 2023-07-05 afbc84f
fixup! Implement `BaseMultiGraph` class
