+++ netgraph/_multigraph.py
@@ -1,0 +1,333 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+"""
+Implements multi-graph classes.
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+
+from ._main import (
+    BaseGraph,
+    EdgeArtist,
+)
+
+from ._edge_layout import (
+    MultiGraphStraightEdgeLayout,
+    MultiGraphCurvedEdgeLayout,
+    MultiGraphBundledEdgeLayout,
+    MultiGraphArcDiagramEdgeLayout,
+)
+
+from ._utils import (
+    _simplify_multigraph,
+)
+
+
+class BaseMultiGraph(BaseGraph):
+    """The MultiGraph base class.
+
+    Parameters
+    ----------
+    edges : list
+        The edges of the graph, with each edge being represented by a (source node ID, target node ID, edge ID) tuple.
+    nodes : list or None, default None
+        List of nodes. Required argument if any node in the graph is unconnected.
+        If None, `nodes` is initialised to the set of the flattened `edges`.
+    node_layout : str or dict, default 'spring'
+        If `node_layout` is a string, the node positions are computed using the indicated method:
+
+        - 'random'       : place nodes in random positions;
+        - 'circular'     : place nodes regularly spaced on a circle;
+        - 'spring'       : place nodes using a force-directed layout (Fruchterman-Reingold algorithm);
+        - 'dot'          : place nodes using the Sugiyama algorithm; the graph should be directed and acyclic;
+        - 'radial'       : place nodes radially using the Sugiyama algorithm; the graph should be directed and acyclic;
+        - 'community'    : place nodes such that nodes belonging to the same community are grouped together;
+        - 'bipartite'    : place nodes regularly spaced on two parallel lines;
+        - 'multipartite' : place nodes regularly spaced on several parallel lines;
+        - 'shell'        : place nodes regularly spaced on concentric circles;
+        - 'geometric'    : place nodes according to the length of the edges between them.
+
+        If `node_layout` is a dict, keys are nodes and values are (x, y) positions.
+    node_layout_kwargs : dict or None, default None
+        Keyword arguments passed to node layout functions.
+        See the documentation of the following functions for a full description of available options:
+
+        - get_random_layout
+        - get_circular_layout
+        - get_fruchterman_reingold_layout
+        - get_sugiyama_layout
+        - get_radial_tree_layout
+        - get_community_layout
+        - get_bipartite_layout
+        - get_multipartite_layout
+        - get_shell_layout
+        - get_geometric_layout
+
+    node_shape : str or dict, default 'o'
+        Node shape.
+        If the type is str, all nodes have the same shape.
+        If the type is dict, maps each node to an individual string representing the shape.
+        The string specification is as for matplotlib.scatter marker, i.e. one of 'so^>v<dph8'.
+    node_size : float or dict, default 3.
+        Node size (radius).
+        If the type is float, all nodes will have the same size.
+        If the type is dict, maps each node to an individual size.
+
+        .. note:: Values are rescaled by BASE_SCALE (1e-2) to be compatible with layout routines in igraph and networkx.
+
+    node_edge_width : float or dict, default 0.5
+        Line width of node marker border.
+        If the type is float, all nodes have the same line width.
+        If the type is dict, maps each node to an individual line width.
+
+        .. note:: Values are rescaled by BASE_SCALE (1e-2) to be compatible with layout routines in igraph and networkx.
+
+    node_color : matplotlib color specification or dict, default 'w'
+        Node color.
+        If the type is a string or RGBA array, all nodes have the same color.
+        If the type is dict, maps each node to an individual color.
+    node_edge_color : matplotlib color specification or dict, default {DEFAULT_COLOR}
[…]
