### 2023-07-05 419525e
Add tests for _multigraph.py
