### 2023-07-06 acfef93
Ensure `NodeArtist.size` always matches radius; remove `NodeArtist.radius`
