+++ netgraph/_main.py
@@ -317,7 +317,7 @@ class BaseGraph(object):
         node_edge_width = _rescale_dict_values(node_edge_width, BASE_SCALE)
         edge_width      = _rescale_dict_values(edge_width, BASE_SCALE)
 
-        self.node_size = node_size # TODO: replace with node_artist.radius
+        self.node_size = node_size
         self.edge_width = edge_width
         self.origin = origin
         self.scale = scale
@@ -482,13 +482,13 @@ class BaseGraph(object):
                     _normalize_numeric_argument(edge_layout_kwargs['selfloop_radius'], selfloops, 'selfloop_radius')
             else:
                 edge_layout_kwargs['selfloop_radius'] = \
-                    {(node, node) : 1.5 * self.node_artists[node].radius for node, _ in selfloops}
+                    {(node, node) : 1.5 * self.node_size[node] for node, _ in selfloops}
 
         if isinstance(edge_layout, str):
             if edge_layout == "straight":
                 edge_layout = StraightEdgeLayout(self.edges, self.node_positions, **edge_layout_kwargs)
             elif edge_layout == "curved":
-                edge_layout_kwargs.setdefault('node_size', {node : artist.radius for node, artist in self.node_artists.items()})
+                edge_layout_kwargs.setdefault('node_size', self.node_size)
                 edge_layout_kwargs.setdefault('origin', self.origin)
                 edge_layout_kwargs.setdefault('scale', self.scale)
                 edge_layout = CurvedEdgeLayout(self.edges, self.node_positions, **edge_layout_kwargs)
