### 2023-07-06 abc1bae
Bugfix: raise ValueError if supplied sparse matrix format is incorrect
