### 2023-07-06 5ef50f6
Remove superfluous _is_directed
