### 2023-07-06 84af35a
Reorder function definitions
