### 2023-07-07 #5372 @paulbrodersen
@rkurchin, it's been a week since your last ping and @ortega2247 doesn't seem to be very active on github at the moment. Do you think we can move forward despite the last missing check-mark?

The outstanding item on his list is in regard to installation instructions, which I did address extensively above:

> > Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
>
> Brief installations instructions are given in the README on github and extensive installation instructions are given in the documentation on ReadTheDocs [here](https://netgraph.readthedocs.io/en/latest/installation.html), both for `pip` and `conda/mamba`. The requirements are documented in various machine readable formats [here](https://github.com/paulbrodersen/netgraph/blob/master/requirements.txt), [here](https://github.com/paulbrodersen/netgraph/blob/master/optional-requirements.txt), [here](https://github.com/paulbrodersen/netgraph/blob/master/pyproject.toml), and [here](https://github.com/paulbrodersen/netgraph/blob/master/setup.py), supporting a wide range of different `pip` versions as well as installation via `conda/mamba` from conda-forge.

Furthermore, in his last comment, he did say

> Looks good to me!

Given his inactivity here since, everything indicates that for him, the review is complete.

I would love to wrap this up before I go on holiday next week, if at all possible.
