### 2023-07-07 702fb46
Bugfix: properly parse multi-graphs as graphs
The previous implementation of `@_handle_multigraph` attempted to sum
edge weights for multi-edges (if any). This was done incorrectly.
The new implementation discard existing edge weights (if any), and
creates a weighted graph, were weights correspond to the number
of multi-edges with the same source and target node.
