### 2023-07-07 2104388
Add `weight_attribute` parameter to `_parse_networkx_graph` and `_parse_igraph_graph`
