+++ netgraph/_parser.py
@@ -220,6 +220,9 @@ def parse_graph(graph):
         raise NotImplementedError("Input graph must be one of: {}\nCurrently, type(graph) = {}".format("\n\n\t" + "\n\t".join(allowed), type(graph)))
 
 
+# --------------------------------------------------------------------------------
+
+
 def is_order_zero(graph):
     """Determine if a graph is an order zero graph, i.e. a graph with no nodes (and no edges)."""
     for check, parser in _check_to_parser.items():
@@ -251,4 +254,166 @@ def is_empty(graph):
             pass
     else:
         allowed = ['list', 'tuple', 'set', 'networkx.Graph', 'igraph.Graph', 'graph_tool.Graph']
+        raise NotImplementedError("Input graph must be one of: {}\nCurrently, type(graph) = {}".format("\n\n\t" + "\n\t".join(allowed), type(graph)))
+
+
+# --------------------------------------------------------------------------------
+# multi-graph parsers
+
+def _parse_multigraph_edge_list(edges):
+    """Ensures that the type of edges is a list, and each edge is a 2-tuple."""
+    # Edge list may be an array, or a list of lists. We want a list of tuples.
+    return [(source, target, eid) for (source, target, eid) in edges]
+
+
+def _parse_multigraph_sparse_matrix_format(adjacency):
+    """Parse graphs given in a sparse format, i.e. edge lists or sparse matrix representations."""
+    rows, columns = np.array(adjacency).shape
+
+    if columns == 3:
+        edges = _parse_multigraph_edge_list(adjacency)
+        nodes = _get_unique_nodes(edges)
+        return nodes, edges, None
+
+    elif columns == 4:
+        edge_weight = {(source, target, eid) : weight for (source, target, eid, weight) in adjacency}
+        edges = list(edge_weight.keys())
+        nodes = _get_unique_nodes(edges)
+
+        # In a sparse adjacency format with integer nodes and float weights,
+        # the type of nodes is promoted to the same type as weights.
+        # If all nodes can safely be demoted to ints, then we probably want to do that.
+        save = True
+        for node in nodes:
+            if not isinstance(_save_cast_float_to_int(node), int):
+                save = False
+                break
+        if save:
+            nodes = [_save_cast_float_to_int(node) for node in nodes]
+            edges = [(_save_cast_float_to_int(source), _save_cast_float_to_int(target), eid) for (source, target, eid) in edges]
+            edge_weight = {(_save_cast_float_to_int(source), _save_cast_float_to_int(target), eid) : weight for (source, target, eid), weight in edge_weight.items()}
+
+        if len(set(edge_weight.values())) > 1:
+            return nodes, edges, edge_weight
+        else:
+            return nodes, edges, None
+
+    else:
+        msg = "Multi-graph specification in sparse matrix format needs to consist of an iterable of tuples of length 3 or 4."
+        msg += "Got iterable of tuples of length {}.".format(columns)
+        raise ValueError(msg)
+
+
+def _parse_multigraph_adjacency_matrix(adjacency):
+    """Parse multi-graphs given in an adjacency matrix format."""
+    sources, targets, edge_ids = np.where(adjacency)
+    edges = list(zip(sources.tolist(), targets.tolist(), edge_ids.tolist()))
+    nodes = list(range(adjacency.shape[0]))
+    edge_weights = {(source, target, eid): adjacency[source, target, eid] for (source, target, eid) in edges}
+
+    if len(set(edge_weights.values())) == 1:
+        return nodes, edges, None
+    else:
+        return nodes, edges, edge_weights
+
+
+def _parse_multigraph_nparray(graph):
[…]
