### 2023-07-07 4ea6a1b
Improve testing of graph parsing functions
