+++ tests/test_parser.py
@@ -5,56 +5,293 @@ Test _parser.py.
 
 import numpy as np
 
-from netgraph._main import Graph
-from netgraph._parser import parse_graph
+from netgraph._parser import (
+    _parse_edge_list,
+    _parse_sparse_matrix_format,
+    _parse_adjacency_matrix,
+    _parse_nparray,
+    _parse_networkx_graph,
+    _parse_igraph_graph,
+    _parse_graph_tool_graph,
+    parse_graph,
+)
 
 
-def test_edge_list():
-    g = Graph([(0, 1)])
-    assert g.nodes == [0, 1]
-    assert g.edges == [(0, 1)]
+def test_parse_edge_list():
+    # 1) plain edge list as iterable of tuples
+    provided_edges = [
+        (0, 0),
+        (0, 1),
+    ]
+    returned_edges = _parse_edge_list(provided_edges)
+    for edge in provided_edges:
+        assert edge in returned_edges
+
+    # 2) edge list as array
+    provided_edges = np.array([
+        (0, 0),
+        (0, 1),
+    ])
+    returned_edges = _parse_edge_list(provided_edges)
+    for edge in provided_edges:
+        assert tuple(edge) in returned_edges
 
 
-def test_edge_list_with_weights():
-    g = Graph([(0, 1, 0.5)])
-    assert g.nodes == [0, 1]
-    assert g.edges == [(0, 1)]
+def test_parse_sparse_matrix_format():
+    # 1) unweighted
+    provided_edges = np.array([
+        (0, 0),
+        (0, 1),
+    ])
+    returned_nodes, returned_edges, returned_weights = \
+        _parse_sparse_matrix_format(provided_edges)
+    for edge in provided_edges:
+        assert tuple(edge) in returned_edges
+    assert {0, 1} == set(returned_nodes)
+    assert returned_weights is None
+
+    # 2) weighted
+    provided_edges = np.array([
+        (0, 0, 0.),
+        (0, 1, 1.),
+    ])
+    returned_nodes, returned_edges, returned_weights = \
+        _parse_sparse_matrix_format(provided_edges)
+    for edge in provided_edges:
+        assert tuple(edge[:2]) in returned_edges
+    assert {0, 1} == set(returned_nodes)
+    assert {(0, 0), (0, 1)} == set(returned_weights.keys())
+    assert {0., 1.} == set(returned_weights.values())
+
+    # 3) multi-graph
+    provided_edges = np.array([
+        (0, 0),
+        (0, 1),
+        (0, 1),
+    ])
+    returned_nodes, returned_edges, returned_weights = \
+        _parse_sparse_matrix_format(provided_edges)
+    for edge in provided_edges:
+        assert tuple(edge) in returned_edges
+    assert {0, 1} == set(returned_nodes)
+    assert {(0, 0), (0, 1)} == set(returned_weights.keys())
+    assert {1, 2} == set(returned_weights.values())
 
 
-def test_sparse_matrix_format():
-    g = Graph(np.array([[0, 1, 0.5]]))
-    assert g.nodes == [0, 1]
-    assert g.edges == [(0, 1)]
+def test_parse_adjacency_matrix():
+    # 1) unweighted
+    provided_adjacency = np.array([
+        [1, 1],
+        [0, 0]
+    ])
+    returned_nodes, returned_edges, returned_weights = \
+        _parse_adjacency_matrix(provided_adjacency)
+    desired_edges = [
+        (0, 0),
+        (0, 1),
+    ]
+    for edge in desired_edges:
+        assert edge in returned_edges
+    assert {0, 1} == set(returned_nodes)
+    assert returned_weights is None
+
+    # 2) weighted
[…]
