### 2023-07-07 1d28072
Implement tests for multi-graph parsing functions
