### 2023-07-07 #5372 @rkurchin
Editorial comments:
- line 40: I suspect you need a comma after "minimisation" since I suspect that would always be applicable?
- 46: missing a period at the end of this sentence
- figure 1, optional changes: I might add some dividing lines as if you just glance at the middle of the figure it's hard to tell immediately which code refers to which figure; also consider rendering code in a `monospace font`
- 54: comma after minimum unnecessary
(This paper is up there in the fewest editorial remarks I've ever needed to give 👏)

Personal comments:
Looks like a great package! I do some work with graphs in my own research; I'll definitely keep this in mind next time I need to make some pretty figures! I also commend you on the "Key Design Decisions" section of the paper – it's some of the most educational and thoughtful exposition I've seen in a JOSS paper that I've gotten to edit.
