### 2023-07-10 #5372 @paulbrodersen
I have corrected the punctuation in line 46 & 54. I believe the absence of a comma in line 40 (`"[...]; ; edge crossing minimisation and edge unbundling where applicable)."`) is correct though, as the `"where applicable"` is meant to extend to the `"edge crossing minimisation"`, as not all node layout algorithms lend themselves for explicit edge crossing minimisation in a post-processing step.

> figure 1, optional changes: I might add some dividing lines as if you just glance at the middle of the figure it's hard to tell immediately which code refers to which figure; also consider rendering code in a monospace font

I agree with both points. I will make a new version of the figure, hopefully by tomorrow.

> Looks like a great package! I do some work with graphs in my own research; I'll definitely keep this in mind next time I need to make some pretty figures! I also commend you on the "Key Design Decisions" section of the paper – it's some of the most educational and thoughtful exposition I've seen in a JOSS paper that I've gotten to edit.

Luckily, nobody can see you blush on the internet. Thank you for the kind words.
