### 2023-07-10 #5372 @idoby
@paulbrodersen Just chiming in to say I agree regarding line 40, I read it the same way you describe it
(and I feel somewhat responsible for the design decisions section, which I think every JOSS paper should include 😛 )
