### 2023-07-10 #5372 @paulbrodersen
> and I feel somewhat responsible for the design decisions section

Rightfully so! Amended the acknowledgements accordingly.

---

# Acknowledgements

We thank GitHub users adleris, Allan L. R. Hansen, chenghuzi, Hamed Mohammadpour, and Pablo for contributing various bug fixes, as well as ortega2247 and idoby for their thoughtful reviews of earlier versions of this manuscript and their inspired suggestions, such as adding a section on "Key Design Decisions".
