### 2023-07-11 bf6b9d7
Add emphasize on click and table on hover features
