+++ netgraph/_main.py
@@ -1852,6 +1852,68 @@ class EmphasizeOnHover(object):
                 self.fig.canvas.draw_idle()
 
 
+class EmphasizeOnClick(object):
+    """Emphasize matplotlib artists when clicking on them by desaturating all other artists."""
+
+    def __init__(self, artist_to_mapping):
+        self.artist_to_mapping = artist_to_mapping
+        self.emphasizeable_artists = self.artist_to_mapping.keys()
+        self.artists = list(self.node_artists.values()) + list(
+            self.edge_artists.values()
+        )
+        keys = list(self.node_artists.keys()) + list(self.edge_artists.keys())
+        self.artist_to_key = dict(zip(self.artists, keys))
+        self.mapping = None
+        self._base_alpha = {artist: artist.get_alpha() for artist in self.artists}
+        self.deemphasized_artists = []
+
+        try:
+            (self.fig,) = set(list(artist.figure for artist in self.artists))
+        except ValueError:
+            raise Exception("All artists have to be on the same figure!")
+
+        try:
+            (self.ax,) = set(list(artist.axes for artist in self.artists))
+        except ValueError:
+            raise Exception("All artists have to be on the same axis!")
+
+        self.fig.canvas.mpl_connect("button_release_event", self._on_release)
+
+    def _add_mapping(self, selected_artist):
+        self.mapping = self.artist_to_mapping[selected_artist]
+        emphasized_artists = []
+        for value in self.mapping:
+            if value in self.node_artists:
+                emphasized_artists.append(self.node_artists[value])
+            elif value in self.edge_artists:
+                emphasized_artists.append(self.edge_artists[value])
+        for artist in self.artists:
+            if artist not in emphasized_artists:
+                artist.set_alpha(self._base_alpha[artist] / 5)
+                self.deemphasized_artists.append(artist)
+
+    def _remove_mapping(self):
+        for artist in self.deemphasized_artists:
+            artist.set_alpha(self._base_alpha[artist])
+        self.deemphasized_artists = []
+        self.mapping = None
+
+    def _on_release(self, event):
+        if event.inaxes == self.ax:
+            for artist in self.emphasizeable_artists:
+                if artist.contains(event)[0]:
+                    if self.mapping:
+                        self._remove_mapping()
+                    else:
+                        self._add_mapping(artist)
+                    self.fig.canvas.draw()
+                    break
+            else:
+                if self.mapping:
+                    self._remove_mapping()
+                self.fig.canvas.draw()
+
+
 class DraggableGraphWithGridMode(DraggableGraph):
     """
     Implements a grid-mode, in which node positions are fixed to a grid.
@@ -2080,6 +2142,128 @@ class EmphasizeOnHoverGraph(Graph, EmphasizeOnHover):
                 self.fig.canvas.draw_idle()
 
 
+class EmphasizeOnClickGraph(Graph, EmphasizeOnClick):
+    """Combines `EmphasizeOnClick` with the `Graph` class such that nodes are emphasized when clicking on them with the mouse.
+
+    Parameters
+    ----------
+    graph : various formats
+        Graph object to plot. Various input formats are supported.
+        In order of precedence:
+
+        - Edge list:
+          Iterable of (source, target) or (source, target, weight) tuples,
+          or equivalent (E, 2) or (E, 3) ndarray, where E is the number of edges.
+        - Adjacency matrix:
+          Full-rank (V, V) ndarray, where V is the number of nodes/vertices.
+          The absence of a connection is indicated by a zero.
+
+          .. note:: If V <= 3, any (2, 2) or (3, 3) matrices will be interpreted as edge lists.**
+
+        - networkx.Graph, igraph.Graph, or graph_tool.Graph object
+
+    mouseover_highlight_mapping : dict or None, default None
[…]
