### 2023-07-11 #5372 @paulbrodersen
@rkurchin, I have made some improvements to the figure.

> figure 1, optional changes: I might add some dividing lines as if you just glance at the middle of the figure it's hard to tell immediately which code refers to which figure; also consider rendering code in a monospace font

1. Dividing lines unfortunately were quite distracting and gave the figure a table-like appearance that did not look professional. Instead, I increased the spacing between panels, added titles, and placed the code snippets below the corresponding panels rather than above. Unfortunately, this increased the size of the figure by quite a bit, such that I had to break it up into two parts.
2. I formatted the code snippets to appear more code-like: I switched to a monospace font, and I added a muted background to mimic the typical formatting of code listings on GitHub, SO, etc. The writing is rather small but increasing the fontsize results in overlaps between code snippets from different panels.

Any further suggestions regarding the figures below?

# Part 1
![gallery_portrait_v5a](https://github.com/openjournals/joss-reviews/assets/8046146/b4444bf8-e8b1-402e-b470-c43a63795855)

# Part 2
![gallery_portrait_v5b](https://github.com/openjournals/joss-reviews/assets/8046146/157c1552-637d-4419-a5bc-8dfb2a012b06)
