### 2023-07-11 d0c75f2
Reworked figure to make it more legible
