### 2023-07-11 6385f49
Prevent figures from being placed on the same line
