### 2023-07-11 c688e14
Small change to figure title
