+++ publication/paper.md
@@ -50,7 +50,7 @@ Netgraph is licensed under the General Public License version 3 (GPLv3). The rep
 
 ![Netgraph's node and edge layouts](gallery_portrait_v5a.png){width=90%}
 
-![Additional key features](gallery_portrait_v5b.png){width=90%}
+![Additional features](gallery_portrait_v5b.png){width=90%}
 
 \pagebreak
 
