### 2023-07-12 #5372 @paulbrodersen
@rkurchin, I can't add check marks to your list above, so I am summarizing my efforts here:

> Double check authors and affiliations (including ORCIDs)

Done.

> Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper.

4.13.1

> Archive the release on Zenodo/figshare/etc and post the DOI here.

[![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8138403.svg)](https://doi.org/10.5281/zenodo.8138403)

> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

Done. I amended the automatically created author list as it included all contributors, not just the author and maintainer (i.e. me).

> Make sure that the license listed for the archive is the same as the software license.

Change the "Open license" default on Zenodo to GPLv3 or later.
