### 2023-07-12 88ad8a1
Correct documentation w.r.t. supported multi-graph data formats
