### 2023-07-12 a073911
Implement `MultiGraph` class
