### 2023-07-12 e525479
Add test for `MultiGraph` class
