+++ tests/test_multigraph.py
@@ -7,21 +7,50 @@ import pytest
 import numpy as np
 import matplotlib.pyplot as plt
 
-from netgraph._multigraph import BaseMultiGraph
+from netgraph._multigraph import (
+    BaseMultiGraph,
+    MultiGraph,
+)
+
 
 @pytest.mark.mpl_image_compare
-def test_multigraph_straight_edge_layout():
+def test_basemultigraph_straight_edge_layout():
     edges = [
+        # single uni-directional edge
+        (0, 0, "a"),
+        # single uni-directional self-loop
         (0, 1, "a"),
-        (0, 1, "b"),
-        (0, 1, "c"),
-        (0, 0, "d"),
-        (0, 0, "e"),
-        (0, 0, "f"),
+        # two uni-directional edges
+        (1, 1, "a"),
+        (1, 1, "b"),
+        # two uni-directional self-lops
+        (1, 2, "a"),
+        (1, 2, "b"),
+        # three
+        (2, 2, "a"),
+        (2, 2, "b"),
+        (2, 2, "c"),
+        (2, 0, "a"),
+        (2, 0, "b"),
+        (2, 0, "c"),
+        # single bi-directional edge
+        (0, 3, "a"),
+        (3, 0, "a"),
+        # two bi-directional edges
+        (1, 3, "a"),
+        (1, 3, "b"),
+        (3, 1, "a"),
+        (3, 1, "b"),
+        # one uni-directional edge with one bi-directional edge
+        (2, 3, "a"),
+        (2, 3, "b"),
+        (3, 2, "a"),
     ]
     node_layout = {
-        0 : np.array([0.25, 0.5]),
-        1 : np.array([0.75, 0.5]),
+        0 : np.array([0.2, 0.2]),
+        1 : np.array([0.8, 0.2]),
+        2 : np.array([0.8, 0.8]),
+        3 : np.array([0.2, 0.8]),
     }
     fig, ax = plt.subplots()
     BaseMultiGraph(
@@ -29,6 +58,7 @@ def test_multigraph_straight_edge_layout():
         node_layout=node_layout,
         edge_layout="straight",
         node_labels=True,
+        edge_width=0.5,
         arrows=True,
         ax=ax,
     )
@@ -36,18 +66,43 @@ def test_multigraph_straight_edge_layout():
 
 
 @pytest.mark.mpl_image_compare
-def test_multigraph_curved_edge_layout():
+def test_basemultigraph_curved_edge_layout():
     edges = [
+        # single uni-directional edge
+        (0, 0, "a"),
+        # single uni-directional self-loop
         (0, 1, "a"),
-        (0, 1, "b"),
-        (0, 1, "c"),
-        (0, 0, "d"),
-        (0, 0, "e"),
-        (0, 0, "f"),
+        # two uni-directional edges
+        (1, 1, "a"),
+        (1, 1, "b"),
+        # two uni-directional self-lops
+        (1, 2, "a"),
+        (1, 2, "b"),
+        # three
+        (2, 2, "a"),
+        (2, 2, "b"),
+        (2, 2, "c"),
+        (2, 0, "a"),
+        (2, 0, "b"),
+        (2, 0, "c"),
+        # single bi-directional edge
+        (0, 3, "a"),
+        (3, 0, "a"),
+        # two bi-directional edges
+        (1, 3, "a"),
+        (1, 3, "b"),
+        (3, 1, "a"),
+        (3, 1, "b"),
+        # one uni-directional edge with one bi-directional edge
+        (2, 3, "a"),
+        (2, 3, "b"),
+        (3, 2, "a"),
     ]
     node_layout = {
[…]
