### 2023-07-12 #5372 @rkurchin
Some tiny nitpicks:
- the version number in the Zenodo and the one here are off by a patch version (I believe that has to be fixed for us to publish)
- your name is listed slightly differently (with middle initials) in the paper than in the Zenodo (I'm not certain if this will have to match exactly or not, but thought I'd point it out)

Otherwise, I think we're ready to go here!
