### 2023-07-13 #5372 @paulbrodersen
> the version number in the Zenodo and the one here are off by a patch version (I believe that has to be fixed for us to publish)

I only increased the version number because Zenodo did not grab the latest release from my github when I first tried to upload it there. Version 4.13.0 and 4.13.1 are otherwise identical. However, I have merged the master branch back into my JOSS branch to update the version number anyway.

>  your name is listed slightly differently (with middle initials) in the paper than in the Zenodo (I'm not certain if this will have to match exactly or not, but thought I'd point it out)

I added my middle initials to the author name of the Zenodo archive.

> Otherwise, I think we're ready to go here!

Yay!
