### 2023-07-13 #5372 @danielskatz
It looks like there is an incorrect ISBN in the bib file (line 43 of https://github.com/paulbrodersen/netgraph/blob/joss/publication/paper.bib seems to be the/a wrong one)
