### 2023-07-14 68c73f6
Remove incorrect (and superfluous) ISBN and ISSN entries; standardize bibtex formatting
