+++ publication/paper.bib
@@ -5,7 +5,7 @@
   editor    = {Varoquaux, Ga{\"{e}}l and Vaught, Travis and Millman, Jarrod},
   pages     = {11--15},
   title     = {{Exploring Network Structure, Dynamics, and Function using NetworkX}},
-  year      = {2008}
+  year      = {2008},
 }
 @article{Csardi:2006,
   title   = {The igraph software package for complex network research},
@@ -15,7 +15,7 @@
   pages   = {1695},
   year    = {2006},
   url     = {http://igraph.sf.net},
-  }
+}
 @article{Peixoto:2014,
   title   = {The graph-tool python library},
   url     = {http://figshare.com/articles/graph_tool/1164194},
@@ -25,40 +25,37 @@
   year    = {2014},
 }
 @inproceedings{Ellson:2002,
-  author    ="Ellson, John and Gansner, Emden and Koutsofios, Lefteris and North, Stephen C. and Woodhull, Gordon",
-  editor    ="Mutzel, Petra and J{\"u}nger, Michael and Leipert, Sebastian",
-  title     ="Graphviz--- Open Source Graph Drawing Tools",
-  booktitle ="Graph Drawing",
-  year      ="2002",
-  publisher ="Springer Berlin Heidelberg",
-  address   ="Berlin, Heidelberg",
-  pages     ="483--484",
-  isbn      ="978-3-540-45848-7",
-  doi       ="10.1007/3-540-45848-4_57"
-  }
+  author    = {Ellson, John and Gansner, Emden and Koutsofios, Lefteris and North, Stephen C. and Woodhull, Gordon},
+  editor    = {Mutzel, Petra and J{\"u}nger, Michael and Leipert, Sebastian},
+  title     = {Graphviz--- Open Source Graph Drawing Tools},
+  booktitle = {Graph Drawing},
+  year      = {2002},
+  publisher = {Springer Berlin Heidelberg},
+  address   = {Berlin, Heidelberg},
+  pages     = {483--484},
+  isbn      = {978-3-540-45848-7},
+  doi       = {10.1007/3-540-45848-4_57},
+}
 @article{Shannon:2003,
   author   = {Shannon, Paul and Markiel, Andrew and Ozier, Owen and Baliga, Nitin S. and Wang, Jonathan T. and Ramage, Daniel and Amin, Nada and Schwikowski, Benno and Ideker, Trey},
   doi      = {10.1101/gr.1239303},
-  file     = {:home/paul/Mendeley Desktop/Paul Shannon et al. - 2003 - Cytoscape A Software Environment for Integrated Models of Biomolecular Interaction Networks.pdf:pdf},
-  isbn     = {1088-9051 /03},
-  issn     = {1088-9051},
   journal  = {Genome Research},
   month    = {nov},
   number   = {11},
   pages    = {2498--2504},
   pmid     = {14597658},
-  title    = {{Cytoscape: A Software Environment for Integrated Models of Biomolecular Interaction Networks}},
+  title    = {Cytoscape: A Software Environment for Integrated Models of Biomolecular Interaction Networks},
   url      = {http://ci.nii.ac.jp/naid/110001910481/ http://genome.cshlp.org/lookup/doi/10.1101/gr.1239303},
   volume   = {13},
   year     = {2003}
-  }
+}
 @paper{Bastian:2009,
 author     = {Mathieu Bastian and Sebastien Heymann and Mathieu Jacomy},
 title      = {Gephi: An Open Source Software for Exploring and Manipulating Networks},
 conference = {International AAAI Conference on Weblogs and Social Media},
 year       = {2009},
[…]
