### 2023-07-14 #5372 @paulbrodersen
Sorry for the issues with the bibliography.

> It looks like there is an incorrect ISBN in the bib file (line 43 of https://github.com/paulbrodersen/netgraph/blob/joss/publication/paper.bib seems to be the/a wrong one)

Removed the offending ISBN (and ISSN) entry.

> line 50 is also overly protected, for what it's worth, but I can also worry about this when you are done and when I proof-read the draft

Removed the second pair of curly braces.

I also did some minor re-formatting of the remaining entries, and removed clearly superfluous keywords (such as the file location on my hard drive).
