### 2023-07-14 d74344a
Reduce coda duplication between `Graph.__init__` and `MultiGraph.__init__`
