### 2023-07-14 3e11cc8
Implement several `_setup_x` methods to reduce code duplication
- `_setup_dragging_clicking_and_selecting`
- `_setup_emphasis`
- `_setup_grid_mode`
- `_setup_annotations`
- `_setup_table_annotations`
