+++ netgraph/_main.py
@@ -1515,14 +1515,15 @@ class DraggableGraph(Graph, DraggableArtists):
     def __init__(self, *args, **kwargs):
         Graph.__init__(self, *args, **kwargs)
         DraggableArtists.__init__(self, self.node_artists.values())
+        self._setup_dragging_clicking_and_selecting()
 
+
+    def _setup_dragging_clicking_and_selecting(self):
         self._draggable_artist_to_node = dict(zip(self.node_artists.values(), self.node_artists.keys()))
-
         self._clickable_artists.extend(list(self.edge_artists.values()))
         self._selectable_artists.extend(list(self.edge_artists.values()))
         self._base_linewidth.update(dict([(artist, artist._lw_data) for artist in self.edge_artists.values()]))
         self._base_edgecolor.update(dict([(artist, artist.get_edgecolor()) for artist in self.edge_artists.values()]))
-
         # # trigger resize of labels when canvas size changes
         # self.fig.canvas.mpl_connect('resize_event', self._on_resize)
 
@@ -1578,9 +1579,11 @@ class DraggableGraphWithGridMode(DraggableGraph):
     """
 
     def __init__(self, *args, **kwargs):
-
         super().__init__(*args, **kwargs)
+        self._setup_grid_mode()
 
+
+    def _setup_grid_mode(self):
         self.grid = False
         self.grid_dx = 0.05 * self.scale[0]
         self.grid_dy = 0.05 * self.scale[1]
@@ -1739,16 +1742,22 @@ class EmphasizeOnHoverGraph(Graph, EmphasizeOnHover):
     def __init__(self, graph, mouseover_highlight_mapping=None, *args, **kwargs):
         Graph.__init__(self, graph, *args, **kwargs)
 
-        artists = list(self.node_artists.values()) + list(self.edge_artists.values())
-        keys = list(self.node_artists.keys()) + list(self.edge_artists.keys())
-        self.artist_to_key = dict(zip(artists, keys))
-        EmphasizeOnHover.__init__(self, artists)
-
-        if mouseover_highlight_mapping is None: # construct default mapping
-            self.mouseover_highlight_mapping = self._get_default_mouseover_highlight_mapping()
-        else: # this includes empty mappings!
+        self._setup_emphasis()
+        if mouseover_highlight_mapping is not None: # this includes empty mappings!
             self._check_mouseover_highlight_mapping(mouseover_highlight_mapping)
             self.mouseover_highlight_mapping = mouseover_highlight_mapping
+
+
+    def _setup_emphasis(self):
+        self.artist_to_key = self._map_artist_to_key()
+        EmphasizeOnHover.__init__(self, list(self.artist_to_key.keys()))
+        self.mouseover_highlight_mapping = self._get_default_mouseover_highlight_mapping()
+
+
+    def _map_artist_to_key(self):
+        artists = list(self.node_artists.values()) + list(self.edge_artists.values())
+        keys = list(self.node_artists.keys()) + list(self.edge_artists.keys())
+        return dict(zip(artists, keys))
 
 
     def _get_default_mouseover_highlight_mapping(self):
@@ -1952,18 +1961,31 @@ class AnnotateOnClickGraph(Graph, AnnotateOnClick):
 
     def __init__(self, *args, **kwargs):
         Graph.__init__(self, *args, **kwargs)
+        self._setup_annotations(*args, **kwargs)
 
+
+    def _setup_annotations(self, *args, **kwargs):
+        if "annotations" in kwargs:
+            artist_to_annotation = self._map_artist_to_annotation(kwargs["annotations"])
+        else:
+            artist_to_annotation = dict()
+
+        if "annotation_fontdict" in kwargs:
[…]
