### 2023-07-14 7428fa1
Implement `DraggableMultiGraph` class
