### 2023-07-14 e4e421d
Implement `DraggableMultiGraphWithGridMode` class
