### 2023-07-14 d34e499
fixup! Implement several `_setup_x` methods to reduce code duplication
