### 2023-07-14 bce27b1
Implement `InteractiveMultiGraph` class
