+++ netgraph/_multigraph.py
@@ -563,7 +563,7 @@ class DraggableMultiGraph(MultiGraph, DraggableGraph, DraggableArtists):
     """Augments `MultiGraph` to support selection and dragging of node artists with the mouse."""
 
     def __init__(self, *args, **kwargs):
-        MultiGraph.__init__(self, *args, **kwargs)
+        super().__init__(*args, **kwargs)
         DraggableArtists.__init__(self, self.node_artists.values())
         self._setup_dragging_clicking_and_selecting()
 
@@ -577,3 +577,267 @@ class DraggableMultiGraphWithGridMode(DraggableMultiGraph, DraggableGraphWithGri
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)
         self._setup_grid_mode()
+
+
+class InteractiveMultiGraph(DraggableMultiGraphWithGridMode, EmphasizeOnHoverGraph, AnnotateOnClickGraph, TableOnClickGraph):
+    """Extends the `MultiGraph` class to support node placement with the mouse, emphasis of graph elements when hovering over them, and toggleable annotations.
+
+    - Nodes can be selected and dragged around with the mouse.
+    - Nodes and edges are emphasized when hovering over them.
+    - Supports additional annotations that can be toggled on and off by clicking on the corresponding node or edge.
+    - These annotations can also be tables.
+
+    Parameters
+    ----------
+    graph : various formats
+        Graph object to plot. Various input formats are supported.
+        In order of precedence:
+
+        - Edge list:
+          Iterable of (source node ID, target node ID, edge key) or
+          (source node ID, target node ID, edge key, weight) tuples,
+          or equivalent (E, 3) or (E, 4) ndarray (where E is the number of edges).
+        - Adjacency matrix:
+          A (V, V, L) ndarray, where V is the number of nodes/vertices, and L is the number of layers.
+          The absence of a connection is indicated by a zero.
+        - networkx.MultiGraph or igraph.Graph object
+
+    node_layout : str or dict, default 'spring'
+        If `node_layout` is a string, the node positions are computed using the indicated method:
+
+        - 'random'       : place nodes in random positions;
+        - 'circular'     : place nodes regularly spaced on a circle;
+        - 'spring'       : place nodes using a force-directed layout (Fruchterman-Reingold algorithm);
+        - 'dot'          : place nodes using the Sugiyama algorithm; the graph should be directed and acyclic;
+        - 'radial'       : place nodes radially using the Sugiyama algorithm; the graph should be directed and acyclic;
+        - 'community'    : place nodes such that nodes belonging to the same community are grouped together;
+        - 'bipartite'    : place nodes regularly spaced on two parallel lines;
+        - 'multipartite' : place nodes regularly spaced on several parallel lines;
+        - 'shell'        : place nodes regularly spaced on concentric circles;
+        - 'geometric'    : place nodes according to the length of the edges between them.
+
+        If `node_layout` is a dict, keys are nodes and values are (x, y) positions.
+    node_layout_kwargs : dict or None, default None
+        Keyword arguments passed to node layout functions.
+        See the documentation of the following functions for a full description of available options:
+
+        - get_random_layout
+        - get_circular_layout
+        - get_fruchterman_reingold_layout
+        - get_sugiyama_layout
+        - get_radial_tree_layout
+        - get_community_layout
+        - get_bipartite_layout
+        - get_multipartite_layout
+        - get_shell_layout
+        - get_geometric_layout
+
+    node_shape : str or dict, default 'o'
+        Node shape.
+        If the type is str, all nodes have the same shape.
+        If the type is dict, maps each node to an individual string representing the shape.
+        The string specification is as for matplotlib.scatter marker, i.e. one of 'so^>v<dph8'.
+    node_size : float or dict, default 3.
+        Node size (radius).
[…]
