### 2023-07-14 42ece36
fixup! Reduce coda duplication between `Graph.__init__` and `MultiGraph.__init__`
