### 2023-07-14 54c3d6d
fixup! Implement `InteractiveMultiGraph` class
