### 2023-07-14 21a6011
Make `_get_default_mouseover_highlight_mapping` multi-graph compatible
