### 2023-07-14 959be2b
fixup! Implement several `_setup_x` methods to reduce code duplication
