### 2023-07-14 9bc48dc
Only initialize `self.fig` and `self.ax` if none already exist
