### 2023-07-14 1f4edd3
Make `redraw_annotations` a method of `AnnotateOnClick`
This reduces code duplication between InteractiveGraph and InteractiveMultiGraph.
