### 2023-07-14 87b8457
fixup! Implement `InteractiveMultiGraph` class
