### 2023-07-14 633f3c3
fixup! Implement several `_setup_x` methods to reduce code duplication
