+++ publication/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{Hagberg:2008,
   address   = {Pasadena, CA USA},
   author    = {Hagberg, Aric A and Schult, Daniel A and Swart, Pieter J},
-  booktitle = {Proceedings of the 7th Python in Science Conference},
+  booktitle = {Proceedings of the 7th {P}ython in Science Conference},
   editor    = {Varoquaux, Ga{\"{e}}l and Vaught, Travis and Millman, Jarrod},
   pages     = {11--15},
   title     = {{Exploring Network Structure, Dynamics, and Function using NetworkX}},
@@ -14,10 +14,10 @@
   volume  = {Complex Systems},
   pages   = {1695},
   year    = {2006},
-  url     = {http://igraph.sf.net},
+  url     = {http://igraph.org},
 }
 @article{Peixoto:2014,
-  title   = {The graph-tool python library},
+  title   = {The graph-tool {P}ython library},
   url     = {http://figshare.com/articles/graph_tool/1164194},
   doi     = {10.6084/m9.figshare.1164194},
   journal = {figshare},
