### 2023-07-18 #5372 @danielskatz
I also have some small changed to suggest, as shown in https://github.com/paulbrodersen/netgraph/pull/72 - please merge this, or let me know what you disagree with.

In addition, I was confused by the word "artist" when I first read it. You might want to introduce this term and briefly explain it in the context of your software when you first use it.
