### 2023-07-18 #5372 @danielskatz
Thanks @idoby - I would still prefer to have it defined/explained, as JOSS papers may be read by a variety of people with a variety of backgrounds and knowledge
