### 2023-07-19 62a4d81
Define the term `Artist` in the context of matplotlib
