### 2023-07-19 #5372 @paulbrodersen
> While I proofread the paper, can you update the title of the Zenodo archive to match that of the paper?

Updated the title of the Zenodo archive.

> I also have some small changes to suggest, as shown in https://github.com/paulbrodersen/netgraph/pull/72 - please merge this, or let me know what you disagree with.

I merged the PR. All the suggestions and corrections looked good to me. Thank you very much!

> In addition, I was confused by the word "artist" when I first read it. You might want to introduce this term and briefly explain it in the context of your software when you first use it.

I amended the first paragraph mentioning the term "artist" to the following ([commit 62a4d81](https://github.com/paulbrodersen/netgraph/commit/62a4d8167fdcc170d6696cd05c190b1f34d31fec)):

*Each visualisation can be customised in various ways. Most parameters can be set using a scalar or string. In this case, the value is applied to all nodes or edges (depending on the parameter). To style each node or each edge differently, supply a dictionary instead. Furthermore, Netgraph's `NodeArtist` and `EdgeArtist` classes, i.e. the Python objects that instruct a renderer to paint nodes and edges onto the canvas, are derived from Matplotlib's `PathPatch` class; similarly, node and edge labels are Matplotlib `Text` instances. Hence all node artists, edge artists, and labels can be manipulated using standard matplotlib syntax after the initial draw.*

> @idoby Maybe it should be capitalized ("Artist") to be clear.

Writing "Matplotlib's `PathPatch` `Artist` class in the paragraph above seemed convoluted. I hence simply named Netgraph's `NodeArtist` and `EdgeArtist` classes, and defined what they do. In this way, people unfamiliar with the term's usage in the context of matplotlib can understand what these two classes do, and people that are familiar with matplotlib will see the term `Artist` highlighted and won't even need to read the subordinate clause to know what is going on.
