### 2023-07-20 d71fc3e
correct the docstring parameter name `tol`
